Subject: Implement Embedded Shop Search and Modularize MainWindow Logic
Description:

Refactoring: Modularized MainWindow.xaml.cs by splitting code into specialized partial classes (Map, Connection, etc.) under Views/MainWindow/ for better maintainability.
Shop Search: Integrated an embedded WebView2 shop search panel with a dedicated toggle button and sleek bottom-edge UI.
Map Enhancements:
Added support for marker rotation (e.g., for helicopters and vehicles).
Implemented interactive shop markers with clustering(Grouping stacked shops) and hover-popups for viewing item offers.
Added high-quality icons for Chinook, Patrol Helicopter, and Deep Sea events.
Settings: Added new options for "Auto Load Shops" and "Hide Console" in the Settings Modal.
UI/UX: Refined the map overlay layout and optimized the logging console height.
Miscellaneous: Updated the project configuration and installer script to include new assets and source files.
