+++ RustPlusDesktop/MainWindow.xaml.cs
@@ -97,6 +97,14 @@ public partial class MainWindow : Window
     // Chinook Chekcer
     private readonly MonumentWatcher _monumentWatcher = new MonumentWatcher();
 
+    private uint? _trackingEntityId; // NEU: ID des Objekts, dem die Kamera folgt
+    private IReadOnlyList<RustPlusClientReal.DynMarker>? _lastMarkers; // Cache für Interaktionen
+
+    public void StopTracking()
+    {
+        _trackingEntityId = null;
+    }
+
     // Camera thumbs: Throttling & "in-flight"-Wächter
     internal readonly HashSet<string> _camBusy = new(StringComparer.OrdinalIgnoreCase);
     private Dictionary<OverlayToolMode, Button> _toolButtons;
@@ -934,6 +942,11 @@ public partial class MainWindow : Window
             return;
         }
 
+        if (ColSidebar != null)
+        {
+            TrackingService.SidebarWidth = ColSidebar.ActualWidth;
+        }
+
         base.OnClosing(e);
     }
 
@@ -948,6 +961,7 @@ public partial class MainWindow : Window
         if (FindName("TxtVersion") is TextBlock txt)
             txt.Text = $"v{AppInfo.VersionShort}";
         InitCameraUi();
+        ApplySettings();
         _selectedMonitor = WinMonitors.All().Count > 0 ? WinMonitors.All()[0] : null;
         AppendLog($"[items-new] baseDir={baseDir}");
         EnsureNewItemDbLoaded();
@@ -1761,102 +1775,6 @@ public partial class MainWindow : Window
         return row;
     }
 
-    // Eine kompakte, einzeilige Offer-Row für die Suchliste (16x16-Icons)
-    private FrameworkElement BuildOfferRowSearchUI(RustPlusClientReal.ShopOrder o, bool compact)
-    {
-        bool outOfStock = o.Stock <= 0;
-
-        var grid = new Grid { Margin = new Thickness(0, 3, 0, 3) };
-        grid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });                       // L-Icon
-        grid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });                       // L-Text
-        grid.ColumnDefinitions.Add(new ColumnDefinition { Width = new GridLength(1, GridUnitType.Star) });  // Spacer
-        grid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });                       // R-Panel
-
-        // left icon (item)
-        var leftIcon = new Image { Width = 16, Height = 16, Margin = new Thickness(0, 0, 6, 0) };
-        BindIcon(leftIcon, o.ItemShortName, o.ItemId);
-        Grid.SetColumn(leftIcon, 0);
-        grid.Children.Add(leftIcon);
-
-        // ---------- LEFT: name + (xN) + stock ----------
-        var leftText = new StackPanel { Orientation = Orientation.Horizontal, VerticalAlignment = VerticalAlignment.Center };
-
-        if (o.Quantity > 1)
-        {
-            leftText.Children.Add(new TextBlock
-            {
-                Text = $"x{o.Quantity} ",
-                Foreground = SearchSubtle,
-                FontWeight = FontWeights.SemiBold,
-                VerticalAlignment = VerticalAlignment.Center
-            });
-        }
-
-        // Voller Name ermitteln …
-        var fullItemName = ResolveItemName(o.ItemId, o.ItemShortName);
-        // … im compact-Modus kürzen
-        var displayItemName = compact ? Shorten(fullItemName, 12) : fullItemName;
-
-        var itemNameTb = new TextBlock
-        {
-            Text = displayItemName,
-            Foreground = SearchText,
-            VerticalAlignment = VerticalAlignment.Center
-        };
-        // Tooltip mit vollem Namen, damit man’s bei Bedarf sieht
-        if (compact && displayItemName != fullItemName)
-            ToolTipService.SetToolTip(itemNameTb, fullItemName);
-
-        leftText.Children.Add(itemNameTb);
-
-        leftText.Children.Add(new TextBlock
-        {
-            Text = $"   ·   Stock {o.Stock}",
-            Foreground = SearchSubtle,
-            Margin = new Thickness(8, 0, 0, 0),
-            VerticalAlignment = VerticalAlignment.Center
-        });
-        Grid.SetColumn(leftText, 1);
-        grid.Children.Add(leftText);
-
-        // ---------- RIGHT: currency + name + amount ----------
-        var right = new StackPanel { Orientation = Orientation.Horizontal, VerticalAlignment = VerticalAlignment.Center };
-        right.Children.Add(new TextBlock { Text = "→  ", Foreground = SearchSubtle, VerticalAlignment = VerticalAlignment.Center });
-
-        var curIcon = new Image { Width = 16, Height = 16, Margin = new Thickness(0, 0, 6, 0) };
-        BindIcon(curIcon, o.CurrencyShortName, o.CurrencyItemId);
-        right.Children.Add(curIcon);
-
-        var fullCurrencyName = ResolveItemName(o.CurrencyItemId, o.CurrencyShortName);
-        var displayCurrencyName = compact ? Shorten(fullCurrencyName, 12) : fullCurrencyName;
-
-        var curNameTb = new TextBlock
-        {
-            Text = displayCurrencyName + " ",
-            Foreground = SearchText,
-            VerticalAlignment = VerticalAlignment.Center
-        };
-        if (compact && displayCurrencyName != fullCurrencyName)
-            ToolTipService.SetToolTip(curNameTb, fullCurrencyName);
-
-        right.Children.Add(curNameTb);
-
-        right.Children.Add(new TextBlock
-        {
-            Text = o.CurrencyAmount.ToString(),
-            Foreground = SearchText,
-            FontWeight = FontWeights.SemiBold,
-            VerticalAlignment = VerticalAlignment.Center
-        });
-
-        if (outOfStock) grid.Opacity = 0.7;
-
-        Grid.SetColumn(right, 3);
-        grid.Children.Add(right);
-
-        return grid;
-    }
-
     // Sichtbarkeit per Checkbox/Toggle
     private bool _showMonuments = true;
 
@@ -2034,53 +1952,6 @@ public partial class MainWindow : Window
         return dot;
     }
 
-
-    private FrameworkElement BuildShopSearchCard(
-    RustPlusClientReal.ShopMarker s,
-    IEnumerable<RustPlusClientReal.ShopOrder> offers,
-    bool compact)
-    {
-        var card = new Border
-        {
-            Background = SearchCardBg,
-            BorderBrush = SearchCardBrd,
-            BorderThickness = new Thickness(1),
-            CornerRadius = new CornerRadius(8),
-            Padding = new Thickness(10),
-            Margin = new Thickness(0, 6, 0, 6)
-        };
-
-        var root = new StackPanel { Orientation = Orientation.Vertical };
-        card.Child = root;
-
-        // header
-        var head = new StackPanel { Orientation = Orientation.Horizontal, Margin = new Thickness(0, 0, 0, 4) };
-        head.Children.Add(new TextBlock
-        {
-            Text = CleanLabel(s.Label) ?? "Shop",
-            Foreground = SearchText,
-            FontWeight = FontWeights.SemiBold,
-            FontSize = compact ? 13 : 14   // optional: mini tweak im compact mode
-        });
-        head.Children.Add(new TextBlock
-        {
-            Text = $"   [{GetGridLabel(s)}]",
-            Foreground = SearchSubtle,
-            Margin = new Thickness(6, 0, 0, 0),
-            VerticalAlignment = VerticalAlignment.Center,
-            FontSize = compact ? 12 : 12
-        });
-        root.Children.Add(head);
-
-        // lines – HIER compact WEITERGEBEN
-        foreach (var o in offers)
-            root.Children.Add(BuildOfferRowSearchUI(o, compact));
-
-        // click → zur Position zentrieren (Zoom unverändert)
-        card.MouseLeftButtonUp += (_, __) => CenterMapOnWorld(s.X, s.Y);
-
-        return card;
-    }
 
     private Grid CreateShopIconwithBadge(string? shortName, int itemId, int qty)
     {
@@ -2374,60 +2245,6 @@ public partial class MainWindow : Window
         double offY = (hostH - contentH * s) * 0.5;
         return (s, offX, offY);
     }
-
-    // Host (Maus) → Scene-VOR-MapTransform (Pivot-Koordinaten für ScaleAt)
-    private Point HostToScenePreTransform(Point hostPt)
-    {
-        var (s, offX, offY) = GetViewboxScaleAndOffset();
-
-        // erst Viewbox "abziehen"
-        var p = new Point(
-            (hostPt.X - offX) / s,
-            (hostPt.Y - offY) / s);
-
-        // dann MapTransform invertieren
-        var m = MapTransform.Matrix;
-        if (m.HasInverse)
-        {
-            m.Invert();
-            p = m.Transform(p);
-        }
-        return p;
-    }
-
-    // Host-Delta → Scene-Delta (damit Panning pixelgenau folgt)
-    private Vector HostDeltaToSceneDelta(Vector dHost)
-    {
-        var (s, _, _) = GetViewboxScaleAndOffset();
-
-        // MapTransform skaliert zusätzlich; Translation findet in Scene-Einheiten statt
-        var m = MapTransform.Matrix;
-        double sx = m.M11, sy = m.M22;
-        if (Math.Abs(sx) < 1e-9) sx = 1;
-        if (Math.Abs(sy) < 1e-9) sy = 1;
-
-        return new Vector(
-            dHost.X / s,
-            dHost.Y / s);
-    }
-    // erkennt "zusammengeklebte" Orders-Zeilen, die kein echter Shop-Name sind
-    private static bool LooksLikeOrdersLabel(string? s)
-    {
-        if (string.IsNullOrWhiteSpace(s)) return false;
-        var t = s.ToLowerInvariant();
-        return t.Contains("item#") || t.Contains("curr#") || t.Contains("→") || t.Contains(";") || t.Contains("stock");
-    }
-    // kürzt Label & filtert Orders-Zeilen raus
-    private static string? CleanLabel(string? raw)
-    {
-        if (string.IsNullOrWhiteSpace(raw)) return null;
-        var s = raw.Trim().Replace('\r', ' ').Replace('\n', ' ');
-        if (LooksLikeOrdersLabel(s)) return null;     // nicht als Name verwenden
-        if (s.Length > 48) s = s.Substring(0, 48) + "…";
-        return s;
-    }
-
-
 
     private async void StatusTimer_Tick(object? sender, EventArgs e)
     {
@@ -3288,589 +3105,12 @@ public partial class MainWindow : Window
         });
     }
 
-    private async Task EnsureWebView2Async()
-    {
-        var dataFolder = IOPath.Combine(
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
-            "RustPlusDesk", "WebView2");
-        Directory.CreateDirectory(dataFolder);
-
-        var env = await CoreWebView2Environment.CreateAsync(userDataFolder: dataFolder);
-        _webView = new WebView2();
-        WebViewHost.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#0B3A4A"));
-        WebViewHost.Children.Add(_webView);
-        Panel.SetZIndex(_webView, 0);           // WebView standardmäßig unten
-
-        await _webView.EnsureCoreWebView2Async(env);
-        _webView.DefaultBackgroundColor = System.Drawing.Color.Transparent;
-        _webView.CoreWebView2.Settings.AreDevToolsEnabled = true;
-
-        // Optional: etwas „normaleren“ UA setzen
-        _webView.CoreWebView2.Settings.UserAgent =
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 " +
-            "(KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36";
-        _webView.NavigationCompleted += WebView_NavigationCompleted;
-    }
-
-    private async void BtnSteamLogin_Click(object sender, RoutedEventArgs e)
-    {
-        try
-        {
-            // Fallback nutzen:
-            var loopback = new SteamOpenIdLoopbackService();
-            var sid = await loopback.SignInAsync(); // ggf. Port anpassen
-            _vm.SteamId64 = sid;
-            TrackingService.SteamId64 = sid; // Persist globally
-            TxtSteamId.Text = sid;
-            AppendLog($"Steam angemeldet (Loopback): {sid}");
-            _vm.Save();
-            HydrateSteamUiFromStorage();   // Label, Button, Avatar aktualisieren
-        }
-        catch (Exception ex)
-        {
-            MessageBox.Show("Steam-Login fehlgeschlagen: " + ex.Message);
-        }
-    }
-    private async Task UpdateServerStatusAsync()
-    {
-        try
-        {
-            if (_rust is RustPlusClientReal real && _vm.Selected?.IsConnected == true)
-            {
-                var st = await real.GetServerStatusAsync();
-                if (st != null)
-                {
-                    // st ist eine Klasseninstanz → direkt auf die Properties zugreifen
-                    _vm.ServerPlayers = (st.Players >= 0 && st.MaxPlayers >= 0)
-                        ? $"{st.Players}/{st.MaxPlayers}" : "–";
-
-                    _vm.ServerQueue = (st.Queue >= 0)
-                        ? st.Queue.ToString()
-                        : "–";
-
-                    _vm.ServerTime = string.IsNullOrWhiteSpace(st.TimeString)
-                        ? "–"
-                        : st.TimeString;
-
-                    //_vm.ServerWipe = st.WipeUtc?.ToLocalTime().ToString("dd.MM.yyyy") ?? "–";
-                }
-            }
-        }
-        catch
-        {
-            // leise weiter – der Poll läuft einfach erneut
-        }
-    }
-    private void WebView_NavigationCompleted(object? sender, CoreWebView2NavigationCompletedEventArgs e)
-    {
-        if (_webView?.Source is null) return;
-        var url = _webView.Source.ToString();
-        if (_steam.TryExtractSteamId64FromReturnUrl(url, out var sid))
-        {
-            _vm.SteamId64 = sid;
-            TxtSteamId.Text = sid;
-            AppendLog($"Steam angemeldet: {sid}");
-
-            // optional: gleich speichern
-            _vm.Save();
-        }
-    }
-
-    private void BtnAddServer_Click(object sender, RoutedEventArgs e)
-    {
-        var host = Microsoft.VisualBasic.Interaction.InputBox("Server IP/Host:", "Server hinzufügen", "127.0.0.1");
-        var portStr = Microsoft.VisualBasic.Interaction.InputBox("Companion-Port:", "Server hinzufügen", "28082");
-        var token = Microsoft.VisualBasic.Interaction.InputBox("Player-Token (Rust+):", "Server hinzufügen", "");
-        var proxy = Microsoft.VisualBasic.Interaction.InputBox("Facepunch-Proxy verwenden? (y/n)", "Server hinzufügen", "n");
-
-        if (int.TryParse(portStr, out var port) && !string.IsNullOrWhiteSpace(host) && !string.IsNullOrWhiteSpace(token))
-        {
-            var prof = new ServerProfile
-            {
-                Name = $"{host}:{port}",
-                Host = host,
-                Port = port,
-                SteamId64 = _vm.SteamId64,
-                PlayerToken = token,
-                UseFacepunchProxy = proxy.Trim().ToLowerInvariant().StartsWith("y")
-            };
-            _vm.AddServer(prof);
-            _vm.Save();
-        }
-        else
-        {
-            MessageBox.Show("Ungültige Eingaben.", "Fehler", MessageBoxButton.OK, MessageBoxImage.Warning);
-        }
-    }
-
-
-
-    private void ListServers_SelectionChanged(object sender, System.Windows.Controls.SelectionChangedEventArgs e)
-    {
-
-        _lastChatTsForCurrentServer = null;
-
-        // Aus der aktuellen Auswahl im VM lesen
-        if (_vm.Selected is { } prof && !string.IsNullOrWhiteSpace(prof.SteamId64))
-            _vm.SteamId64 = prof.SteamId64;
-
-        HydrateSteamUiFromStorage();   // Label/Avatar aktualisieren
-        // absichtlich leer – Binding Selected → CurrentDevices übernimmt das Umschalten
-        RegisterAllHotkeys();
-        ActivateHotkeysForCurrentServer();
-    }
-
     static readonly JsonSerializerOptions JsonOpt = new()
     {
         PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
         WriteIndented = true,
         DefaultIgnoreCondition = System.Text.Json.Serialization.JsonIgnoreCondition.Never
     };
-
-
-    private async Task PollServerStatusLoopAsync(CancellationToken ct)
-    {
-        while (!ct.IsCancellationRequested)
-        {
-            try
-            {
-                if (_rust is RustPlusClientReal real)
-                {
-                    var st = await real.GetServerStatusAsync(ct);
-                    if (st != null)
-                    {
-                        _vm.ServerPlayers = (st.Players >= 0 && st.MaxPlayers >= 0)
-                            ? $"{st.Players}/{st.MaxPlayers}" : "–";
-
-                        _vm.ServerQueue = (st.Queue >= 0) ? st.Queue.ToString() : "–";
-                        _vm.ServerTime = string.IsNullOrWhiteSpace(st.TimeString) ? "–" : st.TimeString;
-                    }
-                }
-            }
-            catch { /* leise weiterversuchen */ }
-
-            try { await Task.Delay(TimeSpan.FromSeconds(10), ct); } catch { }
-        }
-    }
-    private ServerProfile? _connectedProfile;
-    private async Task HardResetAsync(bool reconnect = false)
-    {
-        _connectedProfile = null;
-        // 1) Laufende Polls/Tokens abbrechen
-        try { StopDynPolling(clearKnown: !reconnect); } catch { }
-        try { StopTeamPolling(); } catch { }
-
-        // Falls du eigene CTS für Status hast:
-        try
-        {
-            _statusCts?.Cancel();
-            _statusCts = null;
-        }
-        catch { }
-
-        // 2) Timer stoppen
-        try { _statusTimer?.Stop(); } catch { }
-        try { _shopTimer?.Stop(); _shopTimer = null; } catch { }
-        try { _storageTimer?.Stop(); _shopTimer = null; } catch { }
-
-        // 3) UI-/In-Memory-State leeren
-        try { TeamMembers.Clear(); } catch { }
-        try { _avatarCache.Clear(); } catch { }
-        try { _lastPresence.Clear(); } catch { }
-        try { ClearAllDeathPins(); } catch { }
-        try { ClearAllToggleBusy(); } catch { }
-
-        // Shopspezifisch, wie du es im Connect auch machst
-        try { _lastShops.Clear(); } catch { }
-        try { _shopLifetimes.Clear(); } catch { }
-        try { _knownShopIds.Clear(); } catch { }
-        _initialShopSnapshotTimeUtc = DateTime.MinValue;
-        _alertsNeedRebaseline = true;
-        _lastChatSendUtc = DateTime.MinValue;
-
-        // 4) Overlay-Elemente wirklich vom Canvas runternehmen
-        try
-        {
-            foreach (var el in _shopEls.Values)
-                Overlay.Children.Remove(el);
-            _shopEls.Clear();
-        }
-        catch { }
-
-        // Wenn du noch player-overlays / team-overlays hast:
-        try
-        {
-            ClearUserOverlayElements();   // du hast das im Connect schon – nutzen!
-        }
-        catch { }
-
-        // 5) WIRKLICH vom Rust-Server trennen
-        try
-        {
-            if (_rust != null)
-                await _rust.DisconnectAsync();   // RustPlusClientReal trennt hier sauber und setzt _api = null
-        }
-        catch { }
-
-        // 6) ViewModel „entkoppeln“
-        if (_vm?.Selected != null)
-        {
-            _vm.Selected.IsConnected = false;
-        }
-        if (_vm != null)
-        {
-            _vm.IsBusy = false;
-            _vm.BusyText = "";
-        }
-
-        ResetMapDisplay();
-
-        AppendLog("Hard reset completed.");
-
-        // 7) Optional: direkt wieder verbinden
-        if (reconnect && _vm?.Selected != null)
-        {
-            await Dispatcher.InvokeAsync(async () =>
-            {
-                // wir rufen deine bestehende Logik wieder auf
-                BtnConnect_Click(this, new RoutedEventArgs());
-            });
-        }
-    }
-    private DispatcherTimer? _storageTimer;
-
-    private bool _storageTickBusy; // optionaler Reentrancy-Schutz
-
-
-    private async void BtnHardReset_Click(object sender, RoutedEventArgs e)
-    {
-        var ask = MessageBox.Show(
-            "ARE YOU SURE YOU WANT TO RESET EVERYTHING (WIPE)?\n\n" +
-            "- All servers will be deleted\n" +
-            "- Your Steam Login will be removed\n" +
-            "- The pairing config will be deleted\n\n" +
-            "Continue?", "FULL WIPE DATA", MessageBoxButton.YesNo, MessageBoxImage.Warning);
-
-        if (ask != MessageBoxResult.Yes) return;
-
-        await FullWipeResetAsync();
-    }
-
-    private async Task FullWipeResetAsync()
-    {
-        AppendLog("[WIPE] Starting full reset...");
-        
-        // 1. Connection reset
-        await HardResetAsync(reconnect: false);
-
-        // 2. Clear servers
-        _vm.Servers.Clear();
-        _vm.Save();
-        AppendLog("[WIPE] Servers cleared.");
-
-        // 3. Clear SteamID
-        _vm.SteamId64 = "";
-        TrackingService.SteamId64 = "";
-        AppendLog("[WIPE] Steam credentials removed.");
-
-        // 4. Wipe Pairing Config
-        await ResetPairingConfigAsync(stopListenerFirst: true);
-        AppendLog("[WIPE] Pairing configuration deleted.");
-
-        // 5. Update UI
-        HydrateSteamUiFromStorage();
-        AppendLog("[WIPE] Full wipe completed. Starting new pairing flow...");
-        
-        // Trigger the pairing listener (which will trigger fcm-register because we wiped the config)
-        _ = StartPairingListenerUiAsync();
-    }
-
-    private async void BtnConnect_Click(object sender, RoutedEventArgs e)
-    {
-         await PerformConnectAsync(false);
-    }
-
-    private async void BtnDisconnect_Click(object sender, RoutedEventArgs e)
-    {
-        await HardResetAsync(reconnect: false);
-    }
-
-    private void BtnShowServerInfo_Click(object sender, RoutedEventArgs e)
-    {
-        if (_vm.Selected == null) return;
-        var modal = new Views.ServerInfoModal(_vm.Selected.Name, _vm.Selected.Description) { Owner = this };
-        modal.ShowDialog();
-    }
-
-    private bool _isReconnecting = false;
-    private async void OnConnectionLost()
-    {
-        if (_isReconnecting) return;
-        _isReconnecting = true;
-        
-        AppendLog("[auto-reconnect] Connection lost detected. Starting recovery...");
-        
-        int delay = 2000;
-        int maxDelay = 60000;
-
-        try 
-        {
-            while (_isReconnecting)
-            {
-               AppendLog($"[auto-reconnect] Retrying in {delay/1000}s...");
-               await Task.Delay(delay);
-
-               bool success = await PerformConnectAsync(true);
-               if (success)
-               {
-                   AppendLog("[auto-reconnect] Reconnected successfully!");
-                   _isReconnecting = false;
-                   return;
-               }
-
-               delay = Math.Min(delay * 2, maxDelay);
-            }
-        }
-        catch (Exception ex)
-        {
-             AppendLog("[auto-reconnect] Loop error: " + ex.Message);
-             _isReconnecting = false;
-        }
-    }
-
-    private async Task<bool> PerformConnectAsync(bool silent)
-    {
-        if (!silent)
-        {
-             await HardResetAsync(reconnect: false);
-        // stop polls from previous server
-        _shopTimer?.Stop();
-        _shopTimer = null;
-        StopDynPolling();
-        StopTeamPolling();
-        TeamMembers.Clear();
-        
-        _avatarCache.Clear();
-        _lastPresence.Clear();
-        ClearAllDeathPins();
-        ClearAllToggleBusy();
-        // Shopspezifische State-Tracker leeren:
-        _lastShops.Clear();
-        _shopLifetimes.Clear();
-        _knownShopIds.Clear();
-        _initialShopSnapshotTimeUtc = DateTime.MinValue;
-        // Alerts rebaselinen beim nächsten Poll
-        _alertsNeedRebaseline = true;
-        // Rate-Limiter zurücksetzen
-        _lastChatSendUtc = DateTime.MinValue;
-
-        
-
-
-            // Overlay auf der Map leeren, weil die alten Shop-UI-Elemente vom alten Server sind:
-            foreach (var el in _shopEls.Values)
-                Overlay.Children.Remove(el);
-            _shopEls.Clear();
-        }
-        else
-        {
-             // Silent reconnect: pause pollers but KEEP DATA
-            _shopTimer?.Stop();
-            StopDynPolling(clearKnown: false);
-            StopTeamPolling();
-            _alertsNeedRebaseline = true;
-        }
-        
-
-        //if (_shopTimer =  _shopTimer.Stop();
-        // StopDynPolling();
-
-        if (_vm.Selected is null)
-        {
-            if (!silent) MessageBox.Show("Please chose a server.");
-            return false;
-        }
-
-        try
-        {
-            _vm.IsBusy = true;
-            _vm.BusyText = "Connecting …";
-
-            AppendLog($"Connecting to ws://{_vm.Selected.Host}:{_vm.Selected.Port} …");
-            await _rust.ConnectAsync(_vm.Selected);
-            _vm.Selected.IsConnected = true;
-            AppendLog("Connected.");
-            _connectedProfile = _vm.Selected;
-
-            // Start Battlemetrics tracking polling for this server
-            TrackingService.StartPolling(_vm.Selected.Host ?? "", _vm.Selected.Port, _vm.Selected.Name ?? "");
-
-            // Allow socket to stabilize before firing requests
-            await Task.Delay(1000);
-            // EINMAL casten und überall dieselbe Variable verwenden
-            var real = _rust as RustPlusClientReal;
-            if (real != null)
-            {
-                real.StorageSnapshotReceived -= OnStorageSnapshot;  // doppelte Hooks vermeiden
-                real.StorageSnapshotReceived += OnStorageSnapshot;
-                real.ConnectionLost -= OnConnectionLost;
-                real.ConnectionLost += OnConnectionLost;
-                real.EnsureEventsHooked();
-            }
-
-            // Chat-Marker reset (Deduplication wird via _chatHistoryLog gehandhabt)
-            _lastChatTsForCurrentServer = null;
-
-            // (1) Live-Stream primen (Historie wird beim Öffnen des Chat-Fensters geladen)
-            if (real != null)
-            {
-                try
-                {
-                    await real.PrimeTeamChatAsync(); 
-                }
-                catch (Exception ex)
-                {
-                    AppendLog("[chat] prime error: " + ex.Message);
-                }
-            }
-
-            // (2) Overlay schließen, Map laden, Pairing-Listener starten
-            _vm.IsBusy = false;
-            _vm.BusyText = "";
-            await LoadMapAsync();
-            await StartPairingListenerUiAsync();
-
-            // 3. altes User-Zeug aus vorherigem Server entfernen
-            ClearUserOverlayElements();
-
-            // 4. Team-Overlay-Layer auch leeren (optional aber sinnvoll,
-            //    damit du nicht die PNG vom alten Server siehst)
-            // _visibleOverlayOwners.Clear();
-            // foreach (var img in _playerOverlayElements.Values)
-            //  {
-            //     Overlay.Children.Remove(img);
-            //  }
-            //   _playerOverlayElements.Clear();
-            // 5. eigenes Overlay (falls vorhanden) aus JSON dieses Servers laden
-            // sicherstellen, dass ich sichtbar bin
-            _visibleOverlayOwners.Add(_mySteamId);
-            // von Disk lesen
-            LoadOverlayFromDiskForPlayer(_mySteamId);
-
-            // (3) Geräte in-place rehydrieren (Collection-Instanz behalten)
-            RehydrateDevicesFromStorageInto(_vm.Selected);
-            if (real != null && _vm.Selected?.Devices?.Any() == true)
-            {
-                var storageIds = _vm.Selected.Devices
-                    .Where(d => string.Equals(d.Kind, "StorageMonitor", StringComparison.OrdinalIgnoreCase))
-                    .Select(d => d.EntityId)
-                    .ToList();
-
-                if (storageIds.Count > 0)
-                {
-                    try
-                    {
-                        await real.PrimeSubscriptionsAsync(storageIds);
-                        AppendLog($"PrimeSubscriptions: {storageIds.Count} StorageMonitors.");
-                    }
-                    catch (Exception ex)
-                    {
-                        AppendLog("PrimeSubscriptions Error: " + ex.Message);
-                    }
-                }
-            }
-            await PrimeDeviceKindsAsync(); // <<< add this line
-            _vm.NotifyDevicesChanged();
-            AppendLog($"Devices rehydrated: {_vm.Selected.Devices?.Count ?? 0}");
-
-            // (3b) Kameras rehydrieren und UI darauf umstecken
-            RehydrateCamerasFromStorageInto(_vm.Selected);
-            SwitchCameraSourceTo(_vm.Selected);
-            AppendLog($"Cams rehydrated: {_vm.Selected.CameraIds?.Count ?? 0}");
-
-            // (4) Subscriptions GENAU EINMAL primen (nachdem Devices gesetzt sind)
-            if (real != null && _vm.Selected?.Devices?.Any() == true)
-            {
-                try
-                {
-                    await real.PrimeSubscriptionsAsync(_vm.Selected.Devices.Select(d => d.EntityId));
-                    AppendLog($"PrimeSubscriptions: {_vm.Selected.Devices.Count} IDs.");
-                }
-                catch (Exception ex)
-                {
-                    AppendLog("PrimeSubscriptions Error: " + ex.Message);
-                }
-            }
-
-            // (5) Status-Poll
-            _statusCts?.Cancel();
-            _statusCts = new CancellationTokenSource();
-            _ = PollServerStatusLoopAsync(_statusCts.Token);
-            await UpdateServerStatusAsync();
-            _statusTimer.Start();
-
-            // (6) Team-Polling starten
-            StartTeamPolling();
-            await LoadTeamAsync();
-            if (_overlayToolsVisible)
-            {
-                RebuildOverlayTeamBar();
-            }
-
-        }
-
-        catch (Exception ex)
-        {
-            _vm.IsBusy = false;
-            _vm.BusyText = "";
-            AppendLog("Fehler: " + ex.Message);
-            if (!silent) MessageBox.Show($"Connection failed: {ex.Message}");
-            return false;
-        }
-        _storageTimer?.Stop();
-        _storageTimer = new DispatcherTimer { Interval = TimeSpan.FromSeconds(10) };
-        _storageTimer.Tick += async (_, __) =>
-        {
-            if (_storageTickBusy) return;
-            _storageTickBusy = true;
-            try
-            {
-                var sel = _connectedProfile ?? _vm?.Selected;
-                var devs = sel?.Devices;
-
-                if (devs == null || devs.Count == 0)
-                    return;
-
-                // *** HIER: Snapshot bauen ***
-                var snapshot = devs
-                    .Where(sd => IsStorageDevice(sd))
-                    .ToList();  // Liste von Referenzen, kein Deep Copy
-
-                if (snapshot.Count == 0)
-                    return;
-
-                foreach (var sd in snapshot)
-                {
-                    try
-                    {
-                        await RefreshDeviceStateAsync(sd, log: true, forcePull: true);
-                    }
-                    catch (Exception ex)
-                    {
-                        AppendLog($"[stor/poll] #{sd.EntityId} err: {ex.Message}");
-                    }
-                }
-            }
-            finally
-            {
-                _storageTickBusy = false;
-            }
-        };
-        _storageTimer.Start();
-        // AppendLog("[stor/poll] timer started (10s)");
-        return true;
-    }
-
     // PLAYER DEATH MARKERS AVATAR IMAGE PLAYER DEATH
 
     private bool _showProfileMarkers = true;
@@ -3879,29 +3119,6 @@ public partial class MainWindow : Window
     // death pins per player
     private readonly Dictionary<ulong, FrameworkElement> _deathPins = new();
 
-
-
-
-    private async Task<bool> EnsureConnectedAsync()
-    {
-        if (_vm.Selected is null) { AppendLog("No server selected."); return false; }
-        if (_vm.Selected.IsConnected) return true;
-
-        AppendLog($"Verbinde zu ws://{_vm.Selected.Host}:{_vm.Selected.Port} …");
-        using var cts = new CancellationTokenSource(TimeSpan.FromSeconds(6));
-        try
-        {
-            await _rust.ConnectAsync(_vm.Selected, cts.Token);
-            _vm.Selected.IsConnected = true;
-            AppendLog("Connected.");
-            return true;
-        }
-        catch (Exception ex)
-        {
-            AppendLog("Connect failed: " + ex.Message);
-            return false;
-        }
-    }
     private async void DeviceToggle_Click(object sender, RoutedEventArgs e)
     {
         lock (_toggleBusy)
@@ -4543,200 +3760,6 @@ public partial class MainWindow : Window
         return win.ShowDialog() == true ? result : null;
     }
 
-    private void WebViewHost_MouseWheel(object sender, MouseWheelEventArgs e)
-    {
-        if (_scene == null) return;
-
-        double zoom = e.Delta > 0 ? 1.10 : (1.0 / 1.10);
-        var hostPos = e.GetPosition(WebViewHost);
-
-        ZoomAtHostPosition(hostPos, zoom);
-        e.Handled = true;
-    }
-    private void ZoomAtHostPosition(Point hostPos, double factor)
-    {
-        if (_scene == null) return;
-
-        var pivot = HostToScenePreTransform(hostPos);
-
-        var m = MapTransform.Matrix;
-        m.ScaleAt(factor, factor, pivot.X, pivot.Y);
-        MapTransform.Matrix = m;
-
-        RefreshAllOverlayScales();
-        RefreshMonumentOverlayPositions();
-        RefreshUserOverlayIcons();
-        CenterMiniMapOnPlayer();
-    }
-
-
-    // Welt→Bild (Pixel im Bildkoordinatensystem – vor Zoom/Pan)
-    private const double PAD_WORLD = 2000.0; // exakt der Wert, den du zum Fit benutzt
-
-    private Point WorldToImagePx(double x, double y)
-    {
-        if (_worldSizeS <= 0 || _worldRectPx.Width <= 0 || _worldRectPx.Height <= 0)
-            return new Point(0, 0);
-
-        // das ist genau derselbe "virtuelle" Welt-Rand, den du oben beim Rect benutzt: 1000 pro Seite = 2000 gesamt
-        double totalWorld = _worldSizeS + PAD_WORLD;    // z.B. 4500 + 2000 = 6500
-        double halfPad = PAD_WORLD * 0.5;            // z.B. 1000
-
-        // Rust kann dir X/Y < 0 oder > worldSize schicken → wir kappen nur auf das *erweiterte* Intervall
-        double xx = Math.Clamp(x, -halfPad, _worldSizeS + halfPad);
-        double yy = Math.Clamp(y, -halfPad, _worldSizeS + halfPad);
-
-        // jetzt müssen wir auf das *gesamte* (Welt + Wasser) Quadrat im Bild normalisieren,
-        // nicht nur auf das innere _worldRectPx.
-        //
-        // Aus ComputeWorldRectFromWorldSize wissen wir:
-        // - das innere Weltquadrat hat die Breite _worldRectPx.Width
-        // - dieses innere Quadrat entspricht dem Anteil worldSize / (worldSize + PAD_WORLD)
-        // ⇒ daraus können wir die "volle" Seitenlänge im Bild wiederherleiten:
-        double fullSidePx = _worldRectPx.Width * (totalWorld / _worldSizeS);   // also: innerWidth * (6500 / 4500) z.B.
-
-        // dieses volle Quadrat ist wieder mittig im Bild
-        // (genau wie oben ox/oy berechnet wurden)
-        double imgW = _scene?.Width > 0 ? _scene.Width : ImgMap.Width;
-        double imgH = _scene?.Height > 0 ? _scene.Height : ImgMap.Height;
-        double fullOx = (imgW - fullSidePx) / 2.0;
-        double fullOy = (imgH - fullSidePx) / 2.0;
-
-        // und jetzt ganz normal: verschieben um halfPad, dann auf fullSidePx skalieren
-        double u = fullOx + ((xx + halfPad) / totalWorld) * fullSidePx;
-        double v = fullOy + (((_worldSizeS - yy) + halfPad) / totalWorld) * fullSidePx;
-
-        return new Point(u, v);
-    }
-
-    // Weltkoordinate -> Bildpixel (vor Zoom/Pan), benutzt _worldRectPx/_worldSizeS
-
-
-    // (optional) Rückweg, falls du ihn brauchst
-    private Point ImagePxToWorld(double u, double v)
-    {
-        if (_worldSizeS <= 0 || _worldRectPx.Width <= 0 || _worldRectPx.Height <= 0) return new Point(0, 0);
-
-        double x = (u - _worldRectPx.X) / _worldRectPx.Width * _worldSizeS;
-        double y = _worldSizeS - (v - _worldRectPx.Y) / _worldRectPx.Height * _worldSizeS;
-        return new Point(x, y);
-    }
-
-    private void ChkGrid_Checked(object sender, RoutedEventArgs e) => RedrawGrid();
-
-    private void RedrawGrid()
-    {
-        GridLayer.Children.Clear();
-        if (ChkGrid.IsChecked != true || _worldSizeS <= 0 || _worldRectPx.Width <= 0) return;
-
-        // Zellen: 150 world units
-        int cells = Math.Max(1, (int)Math.Round(_worldSizeS / 150.0));
-
-        double ox = _worldRectPx.X, oy = _worldRectPx.Y;
-        double ow = _worldRectPx.Width, oh = _worldRectPx.Height;
-        double step = ow / cells;
-
-        var stroke = new SolidColorBrush(Color.FromArgb(120, 255, 255, 255));
-        double thin = 1.0, thick = 2.0;
-
-        // senkrecht
-        for (int i = 0; i <= cells; i++)
-        {
-            double x = ox + i * step;
-            var line = new System.Windows.Shapes.Line
-            {
-                X1 = x,
-                Y1 = oy,
-                X2 = x,
-                Y2 = oy + oh,
-                Stroke = stroke,
-                StrokeThickness = (i % 5 == 0) ? thick : thin
-            };
-            GridLayer.Children.Add(line);
-        }
-
-        // waagrecht
-        for (int j = 0; j <= cells; j++)
-        {
-            double y = oy + j * step;
-            var line = new System.Windows.Shapes.Line
-            {
-                X1 = ox,
-                Y1 = y,
-                X2 = ox + ow,
-                Y2 = y,
-                Stroke = stroke,
-                StrokeThickness = (j % 5 == 0) ? thick : thin
-            };
-            GridLayer.Children.Add(line);
-        }
-
-        // Labels (oben-links in jeder Zelle): "A0", "A1", ..., "Z25", "AA...", etc.
-        for (int i = 0; i < cells; i++)
-        {
-            string col = ColumnLabel(i);
-            for (int j = 0; j < cells; j++)
-            {
-                var tb = new TextBlock
-                {
-                    Text = $"{col}{j}",
-                    Foreground = Brushes.White,
-                    FontSize = 10,
-                    Margin = new Thickness(2, 2, 0, 0),
-                    Background = new SolidColorBrush(Color.FromArgb(96, 0, 0, 0)), // dezente Box
-                    Padding = new Thickness(2, 0, 2, 0)
-                };
-
-                double x = ox + i * step + 1; // etwas Abstand von der Linie
-                double y = oy + j * step + 1;
-
-                GridLayer.Children.Add(tb);
-                Canvas.SetLeft(tb, x);
-                Canvas.SetTop(tb, y);
-            }
-        }
-    }
-
-    private static string ColumnLabel(int index)
-    {
-        // 0 -> A, 25 -> Z, 26 -> AA, …
-        var s = "";
-        index++;
-        while (index > 0)
-        {
-            index--;
-            s = (char)('A' + (index % 26)) + s;
-            index /= 26;
-        }
-        return s;
-    }
-
-    private bool TryGetGridRef(double x, double y, out string label)
-    {
-        label = "";
-        if (_worldSizeS <= 0) return false;
-
-        // Off-Grid (Oilrig, Labs) markieren wir als “off-grid”
-        //   if (x < 0 || y < 0 || x > _worldSizeS || y > _worldSizeS)
-        //  {
-        //      label = "off-grid";
-        //      return false;
-        //  }
-
-        // Anzahl Zellen entlang einer Kante – so zeichnest du auch das Grid
-        int cells = Math.Max(1, (int)Math.Round(_worldSizeS / 150.0));
-        double cell = _worldSizeS / (double)cells;
-
-        int col = Math.Clamp((int)Math.Floor(x / cell), 0, cells - 1);
-        int row = Math.Clamp((int)Math.Floor((_worldSizeS - y) / cell), 0, cells - 1);
-
-        label = $"{ColumnLabel(col)}{row}";
-        return true;
-    }
-
-    private string GetGridLabel(RustPlusClientReal.ShopMarker s)
-        => TryGetGridRef(s.X, s.Y, out var g) ? g : "off-grid";
-
     private static string FormatItemName(int id) => /* deine vorhandene Map-Funktion */ ResolveItemName(id, null);
     public static System.Windows.Media.ImageSource? ResolveItemIcon(int itemId, string? shortName, int decodePx = 32)
     {
@@ -4868,318 +3891,18 @@ public partial class MainWindow : Window
         });
     }
 
-    private bool _deepSeaActive = false;
-    private void CheckDeepSeaEvent(IEnumerable<RustPlusClientReal.ShopMarker> shops)
-    {
-        // "Casino Bar Shopkeeper" detection for Deep Sea Event
-        var deepSeaShop = shops.FirstOrDefault(s => 
-            s.Label != null && s.Label.Contains("Casino Bar Shopkeeper")
-        );
-
-        if (deepSeaShop != null)
-        {
-            if (!_deepSeaActive)
-            {
-                _deepSeaActive = true;
-                string dir = GetDeepSeaDirection(deepSeaShop.X, deepSeaShop.Y);
-                // Fire and forget chat
-                _ = SendTeamChatSafeAsync($"Deep Sea Event started! (Direction: {dir})");
-                AppendLog($"[DEEPSEA] Event detected at {deepSeaShop.X:F0},{deepSeaShop.Y:F0} ({dir})");
-            }
-        }
-        else
-        {
-            if (_deepSeaActive)
-            {
-                _deepSeaActive = false;
-                AppendLog("[DEEPSEA] Event ended (shop disappeared).");
-            }
-        }
-    }
-
-    private string GetDeepSeaDirection(double x, double y)
-    {
-        // _worldSizeS is the map size (e.g. 4500)
-        double size = _worldSizeS > 0 ? _worldSizeS : 4500;
-        double margin = 200; // Edge margin
-
-        // Logic based on user request:
-        // Y < 100 (or close to 0) => North
-        // Y > Size - 100 => South
-        // X < 100 => West
-        // X > Size - 100 => East
-
-        if (y < margin) return "North";
-        if (y > size - margin) return "South";
-        if (x < margin) return "West";
-        if (x > size - margin) return "East";
-
-        // Fallback to Grid
-        return TryGetGridRef(x, y, out var g) ? g : "Unknown";
-    }
-
-    private static void PrefetchShopIcons(IEnumerable<RustPlusClientReal.ShopMarker> shops)
-    {
-        foreach (var s in shops)
-        {
-            if (s.Orders == null) continue;
-            foreach (var o in s.Orders)
-            {
-                _ = ResolveItemIcon(o.ItemId, o.ItemShortName);
-                _ = ResolveItemIcon(o.CurrencyItemId, o.CurrencyShortName);
-            }
-        }
-    }
-
     private static Image MakeIcon(string packUri, double size = 32)
     {
-        return new Image
+        var img = new Image
         {
             Width = size,
             Height = size,
             Stretch = Stretch.Uniform,
             Source = new BitmapImage(new Uri(packUri, UriKind.Absolute))
         };
+        RenderOptions.SetBitmapScalingMode(img, BitmapScalingMode.HighQuality);
+        return img;
     }
-    private Point _panLastHost;
-    private Point _lastHost;
-    private void WebViewHost_MouseDown(object? sender, MouseButtonEventArgs e)
-    {
-        WebViewHost.Focus();
-        var hostPos = e.GetPosition(WebViewHost);
-        var mapPos = HostToScenePreTransform(hostPos); // du hast die Helperfunktion schon
-
-        if (_overlayToolsVisible && _currentTool != OverlayToolMode.None)
-        {
-            HandleOverlayMouseDown(e, mapPos);
-            e.Handled = true;
-            return;
-        }
-        if (e.ChangedButton == MouseButton.Middle || e.ChangedButton == MouseButton.Right)
-        {
-            _isPanning = true;
-            _panLastHost = e.GetPosition(WebViewHost);
-            WebViewHost.CaptureMouse();
-            e.Handled = true;
-        }
-    }
-
-    // Handler für Zoom mit NumPad +/-
-    private void WebViewHost_KeyDown(object sender, KeyEventArgs e)
-    {
-        if (_scene == null) return;
-
-        bool zoomIn = e.Key == Key.Add || e.Key == Key.OemPlus;
-        bool zoomOut = e.Key == Key.Subtract || e.Key == Key.OemMinus;
-
-        if (!zoomIn && !zoomOut)
-            return;
-
-        // Nur zoomen, wenn die Maus wirklich über der Map liegt
-        var hostPos = Mouse.GetPosition(WebViewHost);
-        if (hostPos.X < 0 || hostPos.Y < 0 ||
-            hostPos.X > WebViewHost.ActualWidth ||
-            hostPos.Y > WebViewHost.ActualHeight)
-            return;
-
-        double factor = zoomIn ? 1.10 : (1.0 / 1.10);
-
-        ZoomAtHostPosition(hostPos, factor);
-        e.Handled = true;
-    }
-
-    private Point _panLastWorld; // Maus in "Welt"/Scene-Koordinaten (vor der Transform)
-    private void WebViewHost_MouseMove(object? sender, MouseEventArgs e)
-    {
-
-        var hostPos = e.GetPosition(WebViewHost);
-        var mapPos = HostToScenePreTransform(hostPos);
-
-        if (_overlayToolsVisible && _currentTool != OverlayToolMode.None)
-        {
-            HandleOverlayMouseMove(e, mapPos);
-            e.Handled = true;
-            return;
-        }
-        if (!_isPanning) return;
-
-        var hostNow = e.GetPosition(WebViewHost);
-        var dHost = hostNow - _panLastHost;
-        _panLastHost = hostNow;
-
-        // Delta sauber in Scene-Einheiten umrechnen!
-        var dScene = HostDeltaToSceneDelta(dHost);
-
-        var m = MapTransform.Matrix;
-        m.Translate(dScene.X, dScene.Y);
-        MapTransform.Matrix = m;
-        CenterMiniMapOnPlayer();
-        e.Handled = true;
-    }
-
-    private void WebViewHost_MouseUp(object? sender, MouseButtonEventArgs e)
-    {
-        var hostPos = e.GetPosition(WebViewHost);
-        var mapPos = HostToScenePreTransform(hostPos);
-
-        if (_overlayToolsVisible && _currentTool != OverlayToolMode.None)
-        {
-            HandleOverlayMouseUp(e, mapPos);
-            e.Handled = true;
-            return;
-        }
-
-        if (_isPanning && (e.ChangedButton == MouseButton.Middle || e.ChangedButton == MouseButton.Right))
-        {
-            _isPanning = false;
-            WebViewHost.ReleaseMouseCapture();
-            e.Handled = true;
-        }
-    }
-
-    private void BuildMonumentOverlays()
-    {
-        if (Overlay == null || _worldSizeS <= 0 || _worldRectPx.Width <= 0) return;
-
-        // alles neu aufbauen (einmalig nach Map-Load reicht i. d. R.)
-        foreach (var kv in _monEls) Overlay.Children.Remove(kv.Value);
-        _monEls.Clear();
-
-        foreach (var m in _monData)
-        {
-            // world→image
-            var p = WorldToImagePx(m.X, m.Y);
-
-            // Name normalisieren + Variante (A/B/C)
-            var key = NormalizeMonName(m.Name, out var variant);
-            var nice = Beautify(m.Name); // deine vorhandene Beautify-Funktion
-            var tt = string.IsNullOrEmpty(variant) ? nice : $"{nice} ({variant})";
-
-            var fe = MakeMonIcon(key, tt, 28);
-            fe.Tag = m;
-
-            Overlay.Children.Add(fe);
-            Panel.SetZIndex(fe, 800); // unter Dyn-Events (900), aber über Map
-            _monEls[key + "@" + p.X.ToString("0") + "," + p.Y.ToString("0")] = fe;
-
-            ApplyCurrentOverlayScale(fe);
-            Canvas.SetLeft(fe, p.X - 14); // halb Größe
-            Canvas.SetTop(fe, p.Y - 14);
-            fe.Visibility = _showMonuments ? Visibility.Visible : Visibility.Collapsed;
-        }
-    }
-    private void RefreshMonumentOverlayPositions()
-    {
-        if (_monEls.Count == 0) return;
-
-        foreach (var fe in _monEls.Values)
-        {
-            if (fe.Tag is ValueTuple<double, double, string> m)
-            {
-                var p = WorldToImagePx(m.Item1, m.Item2); // Item1 = X, Item2 = Y
-                ApplyMonumentScale(fe);
-                Canvas.SetLeft(fe, p.X - fe.RenderSize.Width / 2);
-                Canvas.SetTop(fe, p.Y - fe.RenderSize.Height / 2);
-                Panel.SetZIndex(fe, 800);
-
-            }
-            else if (fe.Tag != null)
-            {
-                // fallback: dynamic oder anonyme Typen
-                dynamic d = fe.Tag;
-                var p = WorldToImagePx((double)d.X, (double)d.Y);
-                ApplyCurrentOverlayScale(fe);
-                Canvas.SetLeft(fe, p.X - 14);
-                Canvas.SetTop(fe, p.Y - 14);
-                Panel.SetZIndex(fe, 800);
-
-            }
-        }
-    }
-    private void SetupMapScene(BitmapSource bmp)
-    {
-        double wDip = bmp.PixelWidth * (96.0 / bmp.DpiX);
-        double hDip = bmp.PixelHeight * (96.0 / bmp.DpiY);
-
-        const double padPx = 000; // Zeichen-Rand
-
-        ImgMap.Stretch = Stretch.None;
-        ImgMap.HorizontalAlignment = HorizontalAlignment.Left;
-        ImgMap.VerticalAlignment = VerticalAlignment.Top;
-        ImgMap.Width = wDip;
-        ImgMap.Height = hDip;
-
-        GridLayer.Width = wDip;
-        GridLayer.Height = hDip;
-        GridLayer.IsHitTestVisible = false;
-
-        // WICHTIG: Overlay größer machen, aber Map nicht anfassen
-        Overlay.Width = wDip + padPx * 2;
-        Overlay.Height = hDip + padPx * 2;
-        Overlay.IsHitTestVisible = true;
-        Overlay.Background = Brushes.Transparent;
-        EnsureShopsHoverPopup();
-
-        _scene ??= new Grid();
-        _scene.Width = wDip + padPx * 2;
-        _scene.Height = hDip + padPx * 2;
-
-        (ImgMap.Parent as Panel)?.Children.Remove(ImgMap);
-        (GridLayer.Parent as Panel)?.Children.Remove(GridLayer);
-        (Overlay.Parent as Panel)?.Children.Remove(Overlay);
-
-        _scene.Children.Clear();
-
-        // Map bei (padPx, padPx)? → NEIN, jetzt bei (0,0)!
-        _scene.Children.Add(ImgMap); Panel.SetZIndex(ImgMap, 0);
-        _scene.Children.Add(GridLayer); Panel.SetZIndex(GridLayer, 1);
-        _scene.Children.Add(Overlay); Panel.SetZIndex(Overlay, 2);
-
-        _scene.RenderTransform = MapTransform;
-
-        if (_mapView == null)
-        {
-            _mapView = new Viewbox { Stretch = Stretch.Uniform, StretchDirection = StretchDirection.Both };
-            WebViewHost.Children.Add(_mapView);
-            Panel.SetZIndex(_mapView, 0);
-        }
-        _mapView.Child = _scene;
-    }
-    private void ResetMapDisplay()
-    {
-        _mapBaseBmp = null;
-        _mapReady = false;
-
-        ImgMap.Source = null;
-        GridLayer.Children.Clear();
-
-        if (MapPlaceholder != null) MapPlaceholder.Visibility = Visibility.Visible;
-        if (_mapView != null) _mapView.Visibility = Visibility.Collapsed;
-
-        if (_miniMap != null)
-        {
-            _miniMap.Close();
-            _miniMap = null;
-            _miniMapBrush = null;
-        }
-    }
-
-    private void ShowMapBasic(BitmapSource bmp)
-    {
-        if (_webView != null) _webView.Visibility = Visibility.Collapsed;
-
-        _mapBaseBmp = bmp;
-        _mapReady = true;
-        if (MapPlaceholder != null) MapPlaceholder.Visibility = Visibility.Collapsed;
-        if (_mapView != null) _mapView.Visibility = Visibility.Visible;
-        _staticMarkers.Clear();            // << keine Testpunkte
-
-        ImgMap.Source = bmp;               // zunächst nackte Map
-        SetupMapScene(bmp);
-        RedrawGrid();
-    }
-
-    private bool _mapReady;
     private static string Beautify(string s)
     {
         if (string.IsNullOrWhiteSpace(s)) return s;
@@ -5206,69 +3929,6 @@ public partial class MainWindow : Window
 
         return new Rect(ox, oy, sidePx, sidePx);
     }
-    private void EnsureShopsHoverPopup()
-    {
-        if (_shopsHoverPopup != null) return;
-
-        Overlay.Background = Brushes.Transparent;
-
-        _shopsHoverWrap = new WrapPanel
-        {
-            Orientation = Orientation.Horizontal,
-            ItemWidth = SHOP_CARD_WIDTH,
-            ItemHeight = double.NaN,
-            Margin = new Thickness(SHOP_GAP),
-        };
-
-        var border = new Border
-        {
-            Background = PopupBg,    // <<< dunkel statt weiß
-            BorderBrush = PopupBrd,   // dezente Linie
-            BorderThickness = new Thickness(1),
-            CornerRadius = new CornerRadius(10),
-            Padding = new Thickness(6),
-            Child = _shopsHoverWrap,
-            Effect = new System.Windows.Media.Effects.DropShadowEffect
-            {
-                ShadowDepth = 2,
-                BlurRadius = 10,
-                Opacity = 0.5
-            }
-        };
-
-        // Wrap nach N Karten pro Zeile
-        double maxContentWidth = SHOPS_WRAP_COLUMNS * (SHOP_CARD_WIDTH + SHOP_GAP);
-        border.MaxWidth = maxContentWidth + 12 /*Padding*/ + 2 /*Border*/;
-
-        _shopsHoverPopup = new Popup
-        {
-            Placement = PlacementMode.RelativePoint,
-            PlacementTarget = Overlay,
-            StaysOpen = true,
-            AllowsTransparency = true,
-            IsHitTestVisible = false,
-            Child = border
-        };
-
-        Overlay.MouseMove += Overlay_MouseMove_ShowMultiShopCards;
-        Overlay.MouseLeave += (_, __) =>
-        {
-            if (_shopsHoverPopup != null) _shopsHoverPopup.IsOpen = false;
-            EnableSingleShopTooltips(true);
-        };
-    }
-
-    private FrameworkElement? FindShopIconRoot(DependencyObject? d)
-    {
-        while (d != null)
-        {
-            if (d is FrameworkElement fe && _shopIconSet.Contains(fe))
-                return fe;
-            d = VisualTreeHelper.GetParent(d);
-        }
-        return null;
-    }
-
     private static string Shorten(string? s, int max = 10)
     {
         if (string.IsNullOrWhiteSpace(s)) return "";
@@ -5276,149 +3936,6 @@ public partial class MainWindow : Window
         if (s.Length <= max) return s;
         return s.Substring(0, Math.Max(1, max - 1)) + "…";
     }
-
-    private void Overlay_MouseMove_ShowMultiShopCards(object? sender, MouseEventArgs e)
-    {
-        if (_shopsHoverPopup == null || _shopsHoverWrap == null) return;
-
-        var pt = e.GetPosition(Overlay);
-        var hits = new List<FrameworkElement>();
-
-        VisualTreeHelper.HitTest(
-            Overlay,
-            d =>
-            {
-                if (d is UIElement uie && uie.Visibility != Visibility.Visible)
-                    return HitTestFilterBehavior.ContinueSkipSelfAndChildren;
-                return HitTestFilterBehavior.Continue;
-            },
-            r =>
-            {
-                var root = FindShopIconRoot(r.VisualHit);
-                if (root != null && !hits.Contains(root)) hits.Add(root);
-                return HitTestResultBehavior.Continue;
-            },
-            new PointHitTestParameters(pt)
-        );
-
-        if (hits.Count > 1)
-        {
-            // Einzel-Tooltips komplett ausschalten, solange Aggregat offen ist:
-            EnableSingleShopTooltips(false);
-
-            _shopsHoverWrap.Children.Clear();
-
-            foreach (var fe in hits)
-            {
-                if (fe.Tag is RustPlusClientReal.ShopMarker s)
-                {
-                    var offers = s.Orders ?? Enumerable.Empty<RustPlusClientReal.ShopOrder>();
-                    var card = BuildShopSearchCard(s, offers, compact: true);    // deine „volle“ Karte
-                    card.Width = SHOP_CARD_WIDTH;                   // feste Breite fürs Wrappen
-                    ToolTipService.SetIsEnabled(card, false);       // Karte selbst ohne Tooltip
-                    _shopsHoverWrap.Children.Add(card);
-                }
-            }
-
-            _shopsHoverPopup.HorizontalOffset = pt.X + 16;
-            _shopsHoverPopup.VerticalOffset = pt.Y + 16;
-            _shopsHoverPopup.IsOpen = true;
-        }
-        else
-        {
-            // Aggregat zu → Einzel-Tooltips wieder an
-            _shopsHoverPopup.IsOpen = false;
-            EnableSingleShopTooltips(true);
-        }
-    }
-    private FrameworkElement BuildOfferRowSearchUI(RustPlusClientReal.ShopOrder o)
-    => BuildOfferRowSearchUI(o, compact: false);
-    private void EnableSingleShopTooltips(bool on)
-    {
-        foreach (var fe in _shopIconSet)
-            ToolTipService.SetIsEnabled(fe, on);
-    }
-
-    private async Task LoadMapAsync()
-    {
-        if (_rust is not RustPlusClientReal real) return;
-
-        var map = await real.GetMapWithMonumentsAsync();
-        if (map == null) { AppendLog("Map: no data received."); return; }
-
-        await Dispatcher.InvokeAsync(() =>
-        {
-            // 1) Map anzeigen
-            ShowMapBasic(map.Bitmap);
-            SetupMapScene(map.Bitmap);
-            _worldSizeS = map.WorldSize;
-            _worldRectPx = ComputeWorldRectFromWorldSize(map.PixelWidth, map.PixelHeight, _worldSizeS, /*padWorld:*/ 2000);
-            RedrawGrid();
-            Dispatcher.InvokeAsync(() =>
-            {
-                RefreshAllOverlayScales();          // Spieler/Death/Shops skalieren
-                RefreshMonumentOverlayPositions();  // Monuments nach Layout korrekt setzen + skalieren
-            }, System.Windows.Threading.DispatcherPriority.Loaded); // oder Render
-            StartDynPolling();
-
-
-            // Layer-Größen & gleiche Transform wie Bild
-            Overlay.Width = ImgMap.Width;
-            Overlay.Height = ImgMap.Height;
-            GridLayer.Width = ImgMap.Width;
-            GridLayer.Height = ImgMap.Height;
-            //  _scene.RenderTransform = MapTransform;
-            //Overlay.RenderTransform = MapTransform;
-            // GridLayer.RenderTransform = MapTransform;
-
-            // Grid initial zeichnen, je nach Checkbox
-            RedrawGrid();
-
-            // Shops-Timer ggf. starten/stoppen
-            //StartShopPolling();
-
-            int imgW = map.PixelWidth;
-            int imgH = map.PixelHeight;
-            int S = map.WorldSize;
-            _monData = map.Monuments.ToList();
-            BuildMonumentOverlays();
-            // 2) Centered playable square purely from worldSize (no pixel scanning)
-            var worldRectPx = ComputeWorldRectFromWorldSize(imgW, imgH, S, padWorld: 2000);
-            AppendLog($"worldRectPx(fromS)=[{(int)worldRectPx.X},{(int)worldRectPx.Y},{(int)worldRectPx.Width}x{(int)worldRectPx.Height}] img={imgW}x{imgH} S={S}");
-
-            // 3) Monuments (no heuristics, no trimming)
-            var mons = map.Monuments.Where(m => !string.IsNullOrWhiteSpace(m.Name)).ToList();
-
-            // _staticMarkers.Clear();
-
-            foreach (var m in mons)
-            {
-                bool off = (m.X < 0) || (m.Y < 0) || (m.X > S) || (m.Y > S);
-
-                // clamp into world bounds for drawing; still flag off-grid
-                double cx = Math.Clamp(m.X, 0, S);
-                double cy = Math.Clamp(m.Y, 0, S);
-
-                // linear map into the centered square (Y inverted)
-                double u = worldRectPx.X + (cx / S) * worldRectPx.Width;
-                double v = worldRectPx.Y + ((S - cy) / S) * worldRectPx.Height;
-
-                // optional: show off-grid (oilrig/uw-lab) with a tiny outward nudge so it's visibly "outside"
-                if (off)
-                {
-                    const double nudge = 0;
-                    if (m.X < 0) u -= nudge; else if (m.X > S) u += nudge;
-                    if (m.Y < 0) v += nudge; else if (m.Y > S) v -= nudge;
-                }
-
-                //  _staticMarkers.Add((u, v, Beautify(m.Name)));
-            }
-
-            // 4) Einmal rendern
-            // ImgMap.Source = ComposeMapWithMarkers(_mapBaseBmp!);
-        });
-    }
-
     private int _shopAutoSeq = 1; // Fallback-Sequenz, wenn ID fehlt
 
     // stabiler Fallback-Key-Hasher (aus X,Y,Label)
@@ -5452,158 +3969,6 @@ public partial class MainWindow : Window
             return h;
         }
     }
-    private void StartDynPolling()
-    {
-        _dynTimer?.Stop();
-        _dynTimer = new DispatcherTimer { Interval = TimeSpan.FromSeconds(1) };
-        _dynTimer.Tick += async (_, __) => await PollDynMarkersOnceAsync();
-        _dynTimer.Start();
-
-    }
-
-    private void StopDynPolling(bool clearKnown = true)
-    {
-        _dynTimer?.Stop();
-        _dynTimer = null;
-
-        foreach (var kv in _dynEls) Overlay.Children.Remove(kv.Value);
-        _dynEls.Clear();
-        if (clearKnown) _dynKnown.Clear();
-    }
-
-    private void ChkPlayers_Checked(object sender, RoutedEventArgs e)
-    {
-        _showPlayers = (ChkPlayers.IsChecked != false);
-        // just reapply visibility now
-        foreach (var kv in _dynEls)
-        {
-            if (kv.Value.Tag is RustPlusClientReal.DynMarker dm)
-            {
-                if (dm.Type == 1) // player
-                    kv.Value.Visibility = _showPlayers ? Visibility.Visible : Visibility.Collapsed;
-            }
-        }
-    }
-
-    private FrameworkElement BuildEventIconHost(FrameworkElement inner, string? tooltip, int size)
-    {
-        // Host bleibt unskaliert
-        var host = new Grid { Width = size, Height = size, IsHitTestVisible = true };
-        if (tooltip != null) ToolTipService.SetToolTip(host, tooltip);
-
-        host.Children.Add(inner);
-
-        // Inneres wird skaliert (Pivot = Mitte)
-        host.Tag = new PlayerMarkerTag
-        {
-            Radius = size * 0.5,          // für zentrierte Positionierung
-            ScaleExp = SHOP_SIZE_EXP,     // Reaktion wie Shops (passt gut für Events)
-            ScaleBaseMult = SHOP_BASE_MULT,
-            ScaleTarget = inner,
-            ScaleCenterX = size * 0.5,
-            ScaleCenterY = size * 0.5
-        };
-
-        return host;
-    }
-
-    private FrameworkElement BuildEventDot(string tooltip, int size = 14)
-    {
-        var dot = new Ellipse
-        {
-            Width = size,
-            Height = size,
-            Fill = Brushes.Orange,
-            Stroke = Brushes.Black,
-            StrokeThickness = 1.5
-        };
-        return BuildEventIconHost(dot, tooltip, size);
-    }
-
-    private async Task PollDynMarkersOnceAsync()
-    {
-        if (_rust is not RustPlusClientReal real) return;
-        if (_worldSizeS <= 0 || _worldRectPx.Width <= 0) return;
-
-        try
-        {
-            // ---------------------------------------------------------
-            // 1. Statische Monumente laden (Lazy Loading)
-            // ---------------------------------------------------------
-            if (!_monumentWatcher.HasMonuments)
-            {
-                var staticMons = await real.GetStaticMonumentsAsync();
-                if (staticMons != null && staticMons.Count > 0)
-                {
-                    _monumentWatcher.SetMonuments(staticMons);
-                }
-            }
-
-            // ---------------------------------------------------------
-            // 2. Echte Marker holen
-            // ---------------------------------------------------------
-            var list = await real.GetDynamicMapMarkersAsync();
-
-            // ---------------------------------------------------------
-            // 3. Watcher Logik ausführen
-            // ---------------------------------------------------------
-            var virtualMarkers = _monumentWatcher.UpdateAndGetVirtualMarkers(list, _dynKnown);
-
-            // ---------------------------------------------------------
-            // 4. Listen kombinieren
-            // ---------------------------------------------------------
-            var combinedList = new List<RustPlusClientReal.DynMarker>(list.Count + virtualMarkers.Count);
-            combinedList.AddRange(list);
-            combinedList.AddRange(virtualMarkers);
-
-            // --- WIEDERHERGESTELLT: Dein Statistik/Logging Block ---
-            if (list.Count > 0)
-            {
-                var d0 = list[0];
-                // AppendLog($"dyn[0]: type={d0.Type} kind={d0.Kind} xy=({d0.X:0},{d0.Y:0}) label={d0.Label ?? "-"}");
-
-                // Verteilung zählen
-                var cPlayers = list.Count(m => m.Type == 1);
-                var cCargo = list.Count(m => m.Type == 5);
-                var cCrate = list.Count(m => m.Type == 6);
-                var cCH47 = list.Count(m => m.Type == 4);
-                var cPatrol = list.Count(m => m.Type == 8);
-
-                // Falls du das Loggen aktivieren willst:
-                // AppendLog($"dyn: total={list.Count} players={cPlayers} ch47={cCH47} cargo={cCargo} crate={cCrate} patrol={cPatrol}");
-            }
-            // --------------------------------------------------------
-
-            // ---------------------------------------------------------
-            // 5. UI aktualisieren
-            // ---------------------------------------------------------
-            UpdateDynUI(combinedList);
-
-            Dispatcher.InvokeAsync(() => RefreshAllOverlayScales(), System.Windows.Threading.DispatcherPriority.Loaded);
-        }
-        catch
-        {
-            // ignore, next tick
-        }
-    }
-
-    private static uint DynFallbackKey(double x, double y, string? label, int type)
-    {
-        unchecked
-        {
-            uint h = 2166136261;
-            void mix(ulong v) { for (int i = 0; i < 8; i++) { h ^= (byte)(v & 0xFF); h *= 16777619; v >>= 8; } }
-            // Position etwas runden, um Jitter zu vermeiden:
-            double rx = Math.Round(x, 1), ry = Math.Round(y, 1);
-            mix(BitConverter.DoubleToUInt64Bits(rx));
-            mix(BitConverter.DoubleToUInt64Bits(ry));
-            h ^= (byte)type; h *= 16777619;
-            if (!string.IsNullOrEmpty(label))
-                foreach (char c in label) { h ^= (byte)c; h *= 16777619; }
-            if (h == 0) h = 1;
-            return h;
-        }
-    }
     private void BtnDonate_Click(object sender, RoutedEventArgs e)
     {
         try
@@ -5623,183 +3988,6 @@ public partial class MainWindow : Window
 
     private void ChatAnnounce_Checked(object sender, RoutedEventArgs e) => _announceSpawns = true;
     private void ChatAnnounce_Unchecked(object sender, RoutedEventArgs e) => _announceSpawns = false;
-    private void UpdateDynUI(IReadOnlyList<RustPlusClientReal.DynMarker> markers)
-    {
-        if (Overlay == null || _worldSizeS <= 0 || _worldRectPx.Width <= 0) return;
-
-        static uint DynFallbackKey(double x, double y, string? label, int type)
-        {
-            unchecked
-            {
-                uint h = 2166136261;
-                void mix(ulong v) { for (int i = 0; i < 8; i++) { h ^= (byte)(v & 0xFF); h *= 16777619; v >>= 8; } }
-                double rx = Math.Round(x, 1), ry = Math.Round(y, 1);
-                mix(BitConverter.DoubleToUInt64Bits(rx));
-                mix(BitConverter.DoubleToUInt64Bits(ry));
-                h ^= (byte)type; h *= 16777619;
-                if (!string.IsNullOrEmpty(label)) foreach (char c in label) { h ^= (byte)c; h *= 16777619; }
-                if (h == 0) h = 1;
-                return h;
-            }
-        }
-
-        var incoming = new HashSet<uint>();
-
-        foreach (var m in markers)
-        {
-            // Statische Monumente ohne Daten filtern
-            if (m.Type == 0 && m.SteamId == 0) continue;
-
-            bool isPlayer = (m.Type == 1);
-            if (isPlayer && m.SteamId != 0)
-                _lastPlayersBySid[m.SteamId] = (m.X, m.Y, ResolvePlayerName(m));
-
-            if (isPlayer && !_showPlayers) continue;
-
-            bool knownEventType = !isPlayer && sDynIconByType.ContainsKey(m.Type);
-            uint key = m.Id != 0 ? m.Id : DynFallbackKey(m.X, m.Y, m.Label ?? m.Kind, m.Type);
-            incoming.Add(key);
-
-            bool online = false, dead = false;
-            if (_lastPresence.TryGetValue(m.SteamId, out var pr)) { online = pr.Item1; dead = pr.Item2; }
-
-            // Death Marker Check
-            if (_showDeathMarkers)
-            {
-                if (_lastPresence.TryGetValue(m.SteamId, out var prevPresence))
-                {
-                    if (!prevPresence.dead && dead)
-                    {
-                        var vm = TeamMembers.FirstOrDefault(t => t.SteamId == m.SteamId);
-                        if (vm != null) { vm.X = m.X; vm.Y = m.Y; Dispatcher.Invoke(() => PlaceDeathPin(vm)); }
-                        else { Dispatcher.Invoke(() => PlaceOrMoveDeathPin(m.SteamId, m.X, m.Y, ResolvePlayerName(m))); }
-                    }
-                }
-            }
-
-            var nameNow = ResolvePlayerName(m);
-
-            // Prüfen, ob Element schon existiert
-            if (!_dynEls.TryGetValue(key, out var el))
-            {
-                // --- NEUES ELEMENT ERSTELLEN ---
-                try
-                {
-                    if (m.Type == 150) // *** CUSTOM CRATE (Mit Tooltip) ***
-                    {
-                        // 1. Icon laden (crate3.png oder Fallback auf crate.png)
-                        var img = MakeIcon("pack://application:,,,/icons/crate3.png", 48);
-
-                        // 2. Host erstellen
-                        // Wir nutzen m.Label (den Timer) direkt als Tooltip!
-                        var host = BuildEventIconHost(img, m.Label, 48);
-
-                        el = host;
-
-                        // 3. WICHTIG: Sofort registrieren, damit es nicht doppelt gemalt wird!
-                        _dynEls[key] = el;
-                        Overlay.Children.Add(el);
-                        Panel.SetZIndex(el, 2000); // Über allem anderen
-
-                        // 4. Initial Positionieren
-                        var pPos = WorldToImagePx(m.X, m.Y);
-                        Canvas.SetLeft(el, pPos.X - 24);
-                        Canvas.SetTop(el, pPos.Y - 24);
-
-                        ApplyCurrentOverlayScale(el);
-                    }
-                    else if (isPlayer)
-                    {
-                        if (_showProfileMarkers) el = BuildPlayerMarker(m.SteamId, nameNow, online, dead);
-                        else el = BuildPlayerDotMarker(m.SteamId, nameNow, online, dead);
-
-                        _dynEls[key] = el;
-                        Overlay.Children.Add(el);
-                        Panel.SetZIndex(el, 950);
-                        ApplyCurrentOverlayScale(el);
-                    }
-                    else
-                    {
-                        FrameworkElement host;
-                        if (knownEventType)
-                        {
-                            try
-                            {
-                                var img = MakeIcon(sDynIconByType[m.Type], 64);
-                                host = BuildEventIconHost(img, m.Label ?? m.Kind, 64);
-                            }
-                            catch
-                            {
-                                host = BuildEventDot($"{m.Kind} ({m.Type})", 14);
-                            }
-                        }
-                        else host = BuildEventDot($"{m.Kind} ({m.Type})", 14);
-
-                        _dynEls[key] = host;
-                        Overlay.Children.Add(host);
-                        Panel.SetZIndex(host, 920);
-                        ApplyCurrentOverlayScale(host);
-
-                        if (_announceSpawns && !_dynKnown.Contains(key))
-                        {
-                            _dynKnown.Add(key);
-                            var grid = GetGridLabel(m.X, m.Y);
-                            var kind = EventKindText(m.Type);
-                            _ = SendTeamChatSafeAsync($"{kind} spawned in at {grid}");
-                        }
-                        el = host;
-                    }
-                }
-                catch { continue; }
-            }
-            else
-            {
-                // --- UPDATE VORHANDENES ELEMENT ---
-                if (m.Type == 150) // *** TIMER UPDATE (Tooltip) ***
-                {
-                    // Einfach den Tooltip am existierenden Element aktualisieren
-                    if (el is FrameworkElement fe)
-                    {
-                        fe.ToolTip = m.Label; // Setzt z.B. "14:59" als Mouseover-Text
-                    }
-
-                    // Position sicherstellen
-                    var pPos = WorldToImagePx(m.X, m.Y);
-                    Canvas.SetLeft(el, pPos.X - 24);
-                    Canvas.SetTop(el, pPos.Y - 24);
-                    continue; // Fertig mit Crate
-                }
-
-                if (isPlayer) UpdatePlayerMarker(ref el, key, m.SteamId, nameNow, online, dead);
-                else if (el.Tag is not PlayerMarkerTag) el.Tag = m;
-            }
-
-            // Standard Positionierung (für alle anderen)
-            if (m.Type != 150)
-            {
-                var p = WorldToImagePx(m.X, m.Y);
-                if (!(el.Tag is PlayerMarkerTag t && t.IsDeathPin))
-                {
-                    double off = (el.Tag is PlayerMarkerTag t2 && t2.Radius > 0) ? t2.Radius : 5.0;
-                    Canvas.SetLeft(el, p.X - off);
-                    Canvas.SetTop(el, p.Y - off);
-                }
-                if (isPlayer) el.Visibility = _showPlayers ? Visibility.Visible : Visibility.Collapsed;
-            }
-        }
-
-        // Aufräumen (Marker entfernen, die weg sind)
-        CenterMiniMapOnPlayer();
-        var gone = _dynEls.Keys.Where(id => !incoming.Contains(id)).ToList();
-        foreach (var id in gone)
-        {
-            if (_dynEls.TryGetValue(id, out var el))
-            {
-                Overlay.Children.Remove(el);
-                _dynEls.Remove(id);
-            }
-        }
-    }
 
     // avatar size on the map
     private const double PlayerAvatarSize = 24;
@@ -5820,6 +4008,7 @@ public partial class MainWindow : Window
         public FrameworkElement? ScaleTarget { get; set; }
         public double ScaleCenterX { get; set; }  // Pivot X
         public double ScaleCenterY { get; set; }  // Pivot Y
+        public double Rotation { get; set; }      // NEU: für orientierte Marker
     }
 
 
@@ -6283,10 +4472,6 @@ public partial class MainWindow : Window
         }
         catch { /* ignore */ }
     }
-    private string GetGridLabel(RustPlusClientReal.DynMarker m) => GetGridLabel(m.X, m.Y);
-
-    private string GetGridLabel(double x, double y)
-    => TryGetGridRef(x, y, out var g) ? g : "off-grid";
     private static string EventKindText(int type) => type switch
     {
         5 => "Cargo Ship",
@@ -6299,35 +4484,6 @@ public partial class MainWindow : Window
         7 => "Building Blocked",
         _ => "Event"
     };
-
-    private async void ChkShops_Checked(object sender, RoutedEventArgs e)
-    {
-        // nur starten, wenn Map-Kontext steht
-        if (ChkShops.IsChecked == true && _worldSizeS > 0 && _worldRectPx.Width > 0)
-        {
-            _shopTimer?.Stop();
-            _shopTimer = new DispatcherTimer { Interval = TimeSpan.FromSeconds(20) };
-            _shopTimer.Tick += async (_, __) => await PollShopsOnceAsync();
-            _shopTimer.Start();
-            //AppendLog("Shops: Polling an (20s).");
-
-            await PollShopsOnceAsync(); // sofort einmal
-            await Dispatcher.InvokeAsync(() =>
-            {
-                RefreshShopIconScales(); // oder RefreshAllOverlayScales(), wenn du es einheitlich halten willst
-            }, System.Windows.Threading.DispatcherPriority.Loaded);
-        }
-        else
-        {
-            _shopTimer?.Stop();
-            _shopTimer = null;
-
-            // UI leeren (optional)
-            foreach (var kv in _shopEls) Overlay.Children.Remove(kv.Value);
-            _shopEls.Clear();
-            //AppendLog("Shops: Polling aus.");
-        }
-    }
 
     private bool _didInitialOverlayRefresh;
 
@@ -6394,115 +4550,16 @@ public partial class MainWindow : Window
         }
 
         double scale = CalcOverlayScale(eff, exp, baseMult);
+        double rotation = (el.Tag is PlayerMarkerTag ptRot) ? ptRot.Rotation : 0;
 
-        // Pivot exakt setzen – kein RenderTransformOrigin nötig
-        target.RenderTransform = new ScaleTransform(scale, scale)
+        target.RenderTransformOrigin = new Point(0.5, 0.5); // Center rotation/scaling
+        var group = new TransformGroup();
+        group.Children.Add(new ScaleTransform(scale, scale));
+        if (rotation != 0)
         {
-            CenterX = centerX,
-            CenterY = centerY
-        };
-    }
-
-    // auf alle Shops anwenden (z.B. nach einem Zoom)
-    private void RefreshShopIconScales()
-    {
-        double eff = GetEffectiveZoom();
-        double scale = CalcOverlayScale(eff, SHOP_SIZE_EXP, SHOP_BASE_MULT);
-
-        foreach (var fe in _shopEls.Values)
-        {
-            fe.RenderTransformOrigin = new Point(0.5, 0.5);
-            fe.RenderTransform = new ScaleTransform(scale, scale);
+            group.Children.Add(new RotateTransform(rotation));
         }
-    }
-
-    private async Task PollShopsOnceAsync()
-    {
-        if (_rust is not RustPlusClientReal real) return;
-
-        try
-        {
-            var shops = await real.GetVendingShopsAsync();
-
-            UpdateShopsUI(shops);
-            _lastShops = shops;
-
-            // Track lifetimes (für suspicious) und bekannte Shops:
-            UpdateShopLifetimes(shops);
-
-            // Falls wir frisch mit einem neuen Server verbunden haben:
-            if (_alertsNeedRebaseline)
-            {
-                RebaselineAllAlertRulesFromCurrentShops(shops);
-
-                // detect new shops: wir setzen die baseline für new-shops auch neu
-                _initialShopSnapshotTimeUtc = DateTime.UtcNow;
-                _knownShopIds.Clear();
-                foreach (var s in shops)
-                    _knownShopIds.Add(s.Id);
-
-                _alertsNeedRebaseline = false;
-            }
-
-            // Neue Shops melden:
-            await DetectNewShopsAsync(shops);
-
-            // Alerts prüfen:
-            await CheckAlerts(shops);
-
-            // Falls Fenster offen ist und sichtbar -> Liste refreshen
-            if (_shopSearchWin?.IsVisible == true)
-                RefreshShopSearchResults();
-        }
-        catch (Exception ex)
-        {
-            //AppendLog("Shops poll: " + ex.Message);
-        }
-    }
-
-
-    private async Task DetectNewShopsAsync(IReadOnlyList<RustPlusClientReal.ShopMarker> shops)
-    {
-        // Erster erfolgreicher Snapshot?
-        if (_initialShopSnapshotTimeUtc == DateTime.MinValue)
-        {
-            _initialShopSnapshotTimeUtc = DateTime.UtcNow;
-            foreach (var s in shops)
-                _knownShopIds.Add(s.Id);
-            return;
-        }
-
-        if (!_notifyNewShopsToChat)
-            return;
-
-        foreach (var s in shops)
-        {
-            if (_knownShopIds.Contains(s.Id))
-                continue; // schon bekannt
-
-            _knownShopIds.Add(s.Id);
-
-            // kurze Preview bauen (bis zu 2 Orders mit Stock > 0)
-            var preview = s.Orders?
-                .Where(o => o.Stock > 0)
-                .Take(3)
-                .Select(o =>
-                {
-                    var left = ResolveItemName(o.ItemId, o.ItemShortName);
-                    var right = ResolveItemName(o.CurrencyItemId, o.CurrencyShortName);
-                    return $"{left} for {o.CurrencyAmount} {right}";
-                })
-                .ToList();
-
-            string offersShort = (preview != null && preview.Count > 0)
-                ? string.Join(", ", preview)
-                : "no stock";
-
-            string msg =
-                $"New shop {(s.Label ?? "Shop")} [{GetGridLabel(s)}]: {offersShort}";
-            AppendLog($"[{DateTime.Now:HH:mm:ss}] Alert [new shop] {(s.Label ?? "Shop")} [{GetGridLabel(s)}]: {offersShort}");
-            await SendTeamChatSafeAsync(msg);
-        }
+        target.RenderTransform = group;
     }
 
     private class TwoStepFlip
@@ -6863,81 +4920,6 @@ public partial class MainWindow : Window
     }
 
 
-    private void UpdateShopLifetimes(IReadOnlyList<RustPlusClientReal.ShopMarker> shops)
-    {
-        // Step 1: Markiere erstmal alle als "nicht gesehen in diesem Poll"
-        foreach (var kv in _shopLifetimes)
-        {
-            kv.Value.LastSeenUtc = null;
-        }
-
-        // Step 2: Gesehene aktualisieren / neu anlegen
-        foreach (var s in shops)
-        {
-            if (!_shopLifetimes.TryGetValue(s.Id, out var life))
-            {
-                life = new ShopLifetimeInfo
-                {
-                    FirstSeenUtc = DateTime.UtcNow,
-                    LastSnapshot = s
-                };
-                _shopLifetimes[s.Id] = life;
-            }
-
-            life.LastSeenUtc = DateTime.UtcNow;
-            life.LastSnapshot = s;
-        }
-
-        // Step 3: Jetzt checken wir auf "suspicious" für die, die NICHT gesehen wurden
-        if (_notifySuspiciousShops)
-        {
-            foreach (var kv in _shopLifetimes.ToList())
-            {
-                var id = kv.Key;
-                var life = kv.Value;
-
-                // War im letzten Poll da, jetzt nicht gesehen = offline gegangen
-                if (life.LastSeenUtc == null && !life.AnnouncedSuspicious)
-                {
-                    string grid = life.LastSnapshot != null ? GetGridLabel(life.LastSnapshot) : "unknown";
-                    AppendLog($"[dbg] Shop {life.LastSnapshot?.Label ?? "(no label)"} [{grid}] offline after {(DateTime.UtcNow - life.FirstSeenUtc).TotalSeconds:0}s");
-                    var lived = DateTime.UtcNow - life.FirstSeenUtc;
-                    if (lived.TotalSeconds <= 60.0)
-                    {
-                        // Alarm
-                        var snap = life.LastSnapshot;
-                        if (snap != null)
-                        {
-                            var preview = snap.Orders?
-                                .Where(o => o.Stock > 0)
-                                .Take(3)
-                                .Select(o =>
-                                {
-                                    var left = ResolveItemName(o.ItemId, o.ItemShortName);
-                                    var right = ResolveItemName(o.CurrencyItemId, o.CurrencyShortName);
-                                    return $"{left} for {o.CurrencyAmount} {right}";
-                                })
-                                .ToList();
-
-                            string firstFew = (preview != null && preview.Count > 0)
-                                ? string.Join(", ", preview)
-                                : "nothing in stock";
-
-                            string msg =
-                                $"Suspicious shop {(snap.Label ?? "Shop")} " +
-                                $"[{GetGridLabel(snap)}] was online {Math.Round(lived.TotalSeconds)}s, sold {firstFew}";
-
-                            // fire & forget, kein await hier weil wir in Loop sind
-                            _ = SendTeamChatSafeAsync(msg);
-                        }
-
-                        life.AnnouncedSuspicious = true;
-                    }
-                }
-            }
-        }
-    }
-
     private void RunShopAnalysis()
     {
         if (_analysisList == null) return;
@@ -6953,692 +4935,7 @@ public partial class MainWindow : Window
         _analysisList.Visibility = Visibility.Visible;
     }
 
-    private void StopShopPolling()
-    {
-        if (_shopTimer != null)
-        {
-            _shopTimer.Stop();
-            _shopTimer.Tick -= null;
-            _shopTimer = null;
-            AppendLog("Shops: Polling off.");
-        }
-        // optional: bestehende UI leeren
-        foreach (var el in _shopEls.Values) Overlay.Children.Remove(el);
-        _shopEls.Clear();
-    }
-
-
-
-    private void UpdateShopsUI(IReadOnlyList<RustPlusClientReal.ShopMarker> shops)
-    {
-        if (Overlay == null || _worldSizeS <= 0 || _worldRectPx.Width <= 0) return;
-
-        // Index eingehender Marker
-        var incoming = new HashSet<uint>();
-        foreach (var s in shops)
-        {
-
-            incoming.Add(s.Id);
-            var p = WorldToImagePx(s.X, s.Y);
-
-            if (!_shopEls.TryGetValue(s.Id, out var el))
-            {
-                // UI-Element: Punkt + Text nebeneinander
-                // var dot = new Ellipse
-                //  { 
-                var icon = MakeIcon("pack://application:,,,/icons/vending.png", 18);
-                //    Width = 10,
-                //   Height = 10,
-                //    Fill = Brushes.OrangeRed,
-                //    Stroke = Brushes.White,
-                //    StrokeThickness = 2,
-                //    Margin = new Thickness(0, 0, 4, 0)
-                //  };
-
-                var txt = new TextBlock
-                {
-                    // Text = string.IsNullOrWhiteSpace(s.Label) ? "(shop)" : s.Label,
-                    Text = "",
-                    Foreground = Brushes.White,
-                    FontSize = 12,
-                    Margin = new Thickness(6, -2, 0, 0)
-                };
-
-                var sp = new StackPanel { Orientation = Orientation.Horizontal, Tag = s, Cursor = Cursors.Hand };
-                sp.Children.Add(icon);
-                sp.Children.Add(txt);
-
-                // Tooltip mit Angeboten
-                sp.ToolTip = BuildShopTooltip(s);
-
-                // Klick
-                sp.MouseLeftButtonUp += ShopElement_Click;
-
-                _shopEls[s.Id] = sp;
-                Overlay.Children.Add(sp);
-                Panel.SetZIndex(sp, 850);
-                el = sp;
-                _shopIconSet.Add(sp);
-                ApplyCurrentOverlayScale(el);
-            }
-            else
-            {
-                // Daten/Tooltip aktualisieren
-                el.Tag = s;
-                if (el is FrameworkElement fe) fe.ToolTip = BuildShopTooltip(s);
-
-                // (optional) sichtbaren Namen aktualisieren
-                if (el is StackPanel sp && sp.Children.Count >= 2 && sp.Children[1] is TextBlock tb)
-                    tb.Text = "";
-                // tb.Text = string.IsNullOrWhiteSpace(s.Label) ? "(shop)" : s.Label;
-                if (el is FrameworkElement fe2) _shopIconSet.Add(fe2);
-            }
-
-            // Position setzen (in Bild-Pixeln; Overlay liegt in der Szene und bekommt keinen eigenen Transform!)
-            Canvas.SetLeft(el, p.X - 5);
-            Canvas.SetTop(el, p.Y - 5);
-        }
-
-        // Entfernen, was nicht mehr kommt
-        var toRemove = _shopEls.Keys.Where(id => !incoming.Contains(id)).ToList();
-        foreach (var id in toRemove)
-        {
-            if (_shopEls.TryGetValue(id, out var el))
-            {
-                Overlay.Children.Remove(el);
-                _shopEls.Remove(id);
-                if (el is FrameworkElement fe) _shopIconSet.Remove(fe);
-            }
-        }
-        CheckDeepSeaEvent(shops);
-        PrefetchShopIcons(shops);
-    }
-
-    private object BuildShopTooltip(RustPlusClientReal.ShopMarker s)
-    {
-
-        // Container (dunkel, rund, Shadow)
-        var card = new Border
-        {
-            Background = new SolidColorBrush(Color.FromRgb(32, 36, 40)),
-            CornerRadius = new CornerRadius(10),
-            BorderBrush = new SolidColorBrush(Color.FromArgb(40, 255, 255, 255)),
-            BorderThickness = new Thickness(1),
-            Padding = new Thickness(10),
-            Effect = new System.Windows.Media.Effects.DropShadowEffect
-            {
-                BlurRadius = 12,
-                ShadowDepth = 0,
-                Opacity = 0.45,
-                Color = Colors.Black
-            }
-        };
-
-        var root = new StackPanel { Orientation = Orientation.Vertical };
-        card.Child = root;
-
-        // Titelzeile: Name + Grid
-        var title = string.IsNullOrWhiteSpace(s.Label) || LooksLikeOrdersLabel(s.Label) ? "Shop" : s.Label;
-        var header = new StackPanel { Orientation = Orientation.Horizontal, Margin = new Thickness(2, 0, 2, 8) };
-        header.Children.Add(new TextBlock
-        {
-            Text = title,
-            Foreground = Brushes.White,
-            FontWeight = FontWeights.SemiBold,
-            FontSize = 14
-        });
-        header.Children.Add(new TextBlock
-        {
-            Text = "  [" + GetGridLabel(s) + "]",
-            Foreground = new SolidColorBrush(Color.FromArgb(200, 200, 200, 200)),
-            FontSize = 12,
-            Margin = new Thickness(6, 0, 0, 0)
-        });
-        root.Children.Add(header);
-
-        // Orders
-        if (s.Orders is { Count: > 0 })
-        {
-            int shown = 0;
-            foreach (var o in s.Orders)
-            {
-                root.Children.Add(BuildOfferRowUI(o));
-                if (++shown >= 12) { root.Children.Add(new TextBlock { Text = "…", Opacity = 0.7, Foreground = Brushes.White }); break; }
-            }
-        }
-        else
-        {
-            root.Children.Add(new TextBlock
-            {
-                Text = "(No offers created)",
-                Foreground = Brushes.White,
-                Opacity = 0.7
-            });
-        }
-
-        return card; // ToolTip.Content darf ein UIElement sein
-    }
-
-
-
-
-    private Window? _shopSearchWin;
-    private TextBox? _searchTb;
-    private CheckBox? _chkSell;
-    private CheckBox? _chkBuy;
-    private ListBox? _searchList;
     private List<RustPlusClientReal.ShopMarker> _lastShops = new(); // füllen wir beim Polling
-
-    private Style BuildNiceButtonStyle(
-    Color bgNormal,
-    Color bgHover,
-    Color bgPressed,
-    Color borderNormal,
-    Color borderPressed,
-    Color fgNormal,
-    Color fgPressed,
-    CornerRadius corner,
-    Thickness padding)
-    {
-        // Border im ControlTemplate
-        var borderFactory = new FrameworkElementFactory(typeof(Border));
-        borderFactory.Name = "Bd";
-        borderFactory.SetValue(Border.BackgroundProperty, new SolidColorBrush(bgNormal));
-        borderFactory.SetValue(Border.BorderBrushProperty, new SolidColorBrush(borderNormal));
-        borderFactory.SetValue(Border.BorderThicknessProperty, new Thickness(1));
-        borderFactory.SetValue(Border.CornerRadiusProperty, corner);
-
-        var contentFactory = new FrameworkElementFactory(typeof(ContentPresenter));
-        contentFactory.SetValue(ContentPresenter.HorizontalAlignmentProperty, HorizontalAlignment.Center);
-        contentFactory.SetValue(ContentPresenter.VerticalAlignmentProperty, VerticalAlignment.Center);
-        contentFactory.SetValue(Control.PaddingProperty, padding);
-
-        borderFactory.AppendChild(contentFactory);
-
-        // ControlTemplate für Button
-        var template = new ControlTemplate(typeof(Button));
-        template.VisualTree = borderFactory;
-
-        // Trigger: Hover
-        {
-            var triggerHover = new Trigger
-            {
-                Property = Button.IsMouseOverProperty,
-                Value = true
-            };
-            triggerHover.Setters.Add(new Setter(Border.BackgroundProperty, new SolidColorBrush(bgHover), "Bd"));
-            template.Triggers.Add(triggerHover);
-        }
-        // Trigger: Pressed
-        {
-            var triggerPressed = new Trigger
-            {
-                Property = Button.IsPressedProperty,
-                Value = true
-            };
-            triggerPressed.Setters.Add(new Setter(Border.BackgroundProperty, new SolidColorBrush(bgPressed), "Bd"));
-            triggerPressed.Setters.Add(new Setter(Border.BorderBrushProperty, new SolidColorBrush(borderPressed), "Bd"));
-            triggerPressed.Setters.Add(new Setter(Control.ForegroundProperty, new SolidColorBrush(fgPressed)));
-            template.Triggers.Add(triggerPressed);
-        }
-        // Trigger: Disabled
-        {
-            var triggerDisabled = new Trigger
-            {
-                Property = UIElement.IsEnabledProperty,
-                Value = false
-            };
-            triggerDisabled.Setters.Add(new Setter(UIElement.OpacityProperty, 0.5, "Bd"));
-            template.Triggers.Add(triggerDisabled);
-        }
-
-        // Style-Objekt zusammenbauen
-        var style = new Style(typeof(Button));
-        style.Setters.Add(new Setter(Control.ForegroundProperty, new SolidColorBrush(fgNormal)));
-        style.Setters.Add(new Setter(Control.CursorProperty, Cursors.Hand));
-        style.Setters.Add(new Setter(Control.BorderThicknessProperty, new Thickness(0))); // Border machen wir im Template
-        style.Setters.Add(new Setter(Control.TemplateProperty, template));
-
-        return style;
-    }
-    private static void StyleHeaderCheckBox(CheckBox cb, string labelText, bool isInitiallyChecked)
-    {
-        cb.IsChecked = isInitiallyChecked;
-        cb.Cursor = Cursors.Hand;
-        cb.ToolTip = labelText;
-        cb.Margin = new Thickness(4, 0, 8, 0);
-        cb.VerticalAlignment = VerticalAlignment.Center;
-        cb.Foreground = Brushes.White;
-        cb.Background = Brushes.Transparent;
-        cb.BorderThickness = new Thickness(0);
-        cb.Focusable = false;
-
-        // Wir bauen das Template einmal hier inline:
-        var circleBorder = new FrameworkElementFactory(typeof(Border));
-        circleBorder.SetValue(Border.WidthProperty, 16.0);
-        circleBorder.SetValue(Border.HeightProperty, 16.0);
-        circleBorder.SetValue(Border.CornerRadiusProperty, new CornerRadius(8));
-        circleBorder.SetValue(Border.BorderThicknessProperty, new Thickness(1));
-        circleBorder.SetValue(Border.MarginProperty, new Thickness(0, 0, 4, 0));
-        circleBorder.SetValue(Border.HorizontalAlignmentProperty, HorizontalAlignment.Left);
-        circleBorder.SetValue(Border.VerticalAlignmentProperty, VerticalAlignment.Center);
-
-        // BorderBrush + Background kommen über Trigger
-        circleBorder.SetValue(Border.BorderBrushProperty,
-            new SolidColorBrush(Color.FromArgb(160, 255, 255, 255))); // default
-        circleBorder.SetValue(Border.BackgroundProperty,
-            new SolidColorBrush(Color.FromRgb(40, 44, 48))); // default dark bg
-
-        // innerer Punkt
-        var dot = new FrameworkElementFactory(typeof(Ellipse));
-        dot.SetValue(Ellipse.WidthProperty, 8.0);
-        dot.SetValue(Ellipse.HeightProperty, 8.0);
-        dot.SetValue(Ellipse.FillProperty,
-            new SolidColorBrush(Color.FromRgb(0, 200, 255))); // cyan-ish
-        dot.SetValue(Ellipse.HorizontalAlignmentProperty, HorizontalAlignment.Center);
-        dot.SetValue(Ellipse.VerticalAlignmentProperty, VerticalAlignment.Center);
-
-        // dot ist nur sichtbar wenn checked -> Trigger unten
-        dot.SetValue(UIElement.VisibilityProperty, Visibility.Collapsed);
-
-        // Border enthält den Punkt
-        var borderPanel = new FrameworkElementFactory(typeof(Grid));
-        borderPanel.AppendChild(dot);
-        circleBorder.AppendChild(borderPanel);
-
-        // daneben der Text ("Sells", "Buys", etc.)
-        var text = new FrameworkElementFactory(typeof(TextBlock));
-        text.SetValue(TextBlock.TextProperty, labelText);
-        text.SetValue(TextBlock.ForegroundProperty, Brushes.White);
-        text.SetValue(TextBlock.FontSizeProperty, 12.0);
-        text.SetValue(TextBlock.VerticalAlignmentProperty, VerticalAlignment.Center);
-
-        // horizontal zusammenbauen
-        var rootPanel = new FrameworkElementFactory(typeof(StackPanel));
-        rootPanel.SetValue(StackPanel.OrientationProperty, Orientation.Horizontal);
-        rootPanel.SetValue(StackPanel.VerticalAlignmentProperty, VerticalAlignment.Center);
-        rootPanel.AppendChild(circleBorder);
-        rootPanel.AppendChild(text);
-
-        // Template definieren
-        var tpl = new ControlTemplate(typeof(CheckBox));
-        tpl.VisualTree = rootPanel;
-
-        // Trigger: wenn Checked == true, dann Border in aktiv-Farbe und Punkt sichtbar
-        var tIsChecked = new Trigger
-        {
-            Property = ToggleButton.IsCheckedProperty,
-            Value = true
-        };
-        tIsChecked.Setters.Add(new Setter(Border.BorderBrushProperty,
-            new SolidColorBrush(Color.FromRgb(0, 200, 255)), "circle"));
-        tIsChecked.Setters.Add(new Setter(Ellipse.VisibilityProperty,
-            Visibility.Visible, dot.Name ?? ""));
-        // Trick: wir müssen den Childs Namen geben, damit wir sie im Setter ansprechen können
-
-        // => wir geben denen Namen:
-        circleBorder.Name = "circle";
-        dot.Name = "dot";
-
-        // jetzt nochmal Trigger definieren MIT Names:
-        tIsChecked = new Trigger
-        {
-            Property = ToggleButton.IsCheckedProperty,
-            Value = true
-        };
-        tIsChecked.Setters.Add(new Setter(Border.BorderBrushProperty,
-            new SolidColorBrush(Color.FromRgb(0, 200, 255)), "circle"));
-        tIsChecked.Setters.Add(new Setter(Border.BackgroundProperty,
-            new SolidColorBrush(Color.FromRgb(20, 30, 36)), "circle")); // leicht blauer bg
-        tIsChecked.Setters.Add(new Setter(UIElement.VisibilityProperty,
-            Visibility.Visible, "dot"));
-
-        tpl.Triggers.Add(tIsChecked);
-
-        cb.Template = tpl;
-    }
-
-    private void BtnShopSearch_Click(object sender, RoutedEventArgs e)
-    {
-        if (_shopSearchWin == null) CreateShopSearchWindow();
-        _shopSearchWin.Show();
-        _shopSearchWin.Activate();
-        RefreshShopSearchResults(); // später: Live-Filter
-        _shopSearchWin.Closed += (sender, args) => _shopSearchWin = null;
-    }
-
-
-
-    private void CreateShopSearchWindow()
-    {
-
-        // Farb-Palette ähnlich deinem globalen Theme
-        Color colBgNormal = Color.FromRgb(30, 87, 111);   // SurfaceAlt
-        Color colBgHover = Color.FromRgb(42, 48, 52);   // Surface
-        Color colBgPressed = Color.FromRgb(10, 58, 74);   // AccentDark-ish (leicht blaugrün)
-        Color colBorderNorm = Color.FromRgb(79, 195, 247); // helle dünne Linie
-        Color colBorderDown = Color.FromRgb(10, 58, 74);   // gleich wie Pressed BG = "eingedrückt"
-        Color colFgNormal = Colors.White;                // TextPrimary
-        Color colFgPressed = Colors.White;
-
-        var corner = new CornerRadius(10); // dein Radius
-        var pad = new Thickness(8, 4, 8, 4);
-
-        Style niceBtnStyle = BuildNiceButtonStyle(
-            colBgNormal,
-            colBgHover,
-            colBgPressed,
-            colBorderNorm,
-            colBorderDown,
-            colFgNormal,
-            colFgPressed,
-            corner,
-            pad
-        );
-
-        var w = new Window
-        {
-            Title = "Shop Search",
-            Width = 560,
-            Height = 560,
-            Owner = this,
-            Background = SearchWinBg,
-            Foreground = SearchText
-        };
-
-        // Root -> vertikal
-        var root = new DockPanel
-        {
-            Margin = new Thickness(10)
-        };
-
-        // ========== OBERER BLOCK (Suchleiste, Buttons) ==========
-        // Wir machen hier einen StackPanel vertical, damit wir 2 Zeilen + Alerts + Filterzeile etc. haben
-        var headerWrap = new StackPanel
-        {
-            Orientation = Orientation.Vertical,
-            Margin = new Thickness(0, 0, 0, 8)
-        };
-
-        // --- Zeile 1: [TextBox][Search][Analyze] ---
-        var row1 = new StackPanel
-        {
-            Orientation = Orientation.Horizontal,
-            Margin = new Thickness(0, 0, 0, 4)
-        };
-
-      
-
-        // Icon links
-        // Farben im Stil deiner UI
-        var colOuterBg = Color.FromRgb(24, 26, 28);        // Gesamtfeld-Hintergrund (dunkel)
-        var colIconBg = Color.FromRgb(18, 20, 22);        // extra dunkler Block hinter dem Icon
-        var colBorder = Color.FromArgb(160, 0, 173, 239); // dein Cyan-ish Border (kannst du anpassen)
-        var colText = Colors.White;
-
-        // Äußere Border = komplette Suchleiste mit runden Ecken
-        var searchOuter = new Border
-        {
-            Background = new SolidColorBrush(colOuterBg),
-            BorderBrush = new SolidColorBrush(colBorder),
-            BorderThickness = new Thickness(1),
-            CornerRadius = new CornerRadius(6),
-            Margin = new Thickness(0, 0, 8, 0), // Abstand rechts zu "Profit Trades"
-            VerticalAlignment = VerticalAlignment.Center,
-            SnapsToDevicePixels = true
-        };
-
-        // Innenlayout: 2 Spalten (Icon | TextBox)
-        var innerGrid = new Grid
-        {
-            VerticalAlignment = VerticalAlignment.Center
-        };
-        innerGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });
-        innerGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = new GridLength(1, GridUnitType.Star) });
-
-        // Icon-Panel links
-        var iconHost = new Border
-        {
-            Background = new SolidColorBrush(colIconBg),
-            BorderBrush = Brushes.Transparent,
-            BorderThickness = new Thickness(0),
-            Padding = new Thickness(6, 4, 6, 4),
-            VerticalAlignment = VerticalAlignment.Center,
-            Child = new TextBlock
-            {
-                Text = "🔍",
-                Foreground = new SolidColorBrush(colText),
-                FontSize = 12,
-                VerticalAlignment = VerticalAlignment.Center
-            }
-        };
-        Grid.SetColumn(iconHost, 0);
-        innerGrid.Children.Add(iconHost);
-
-        // TextBox rechts – ohne eigenen Border, damit’s wie ein Control wirkt
-        _searchTb = new TextBox
-        {
-            Width = 260,
-            Background = Brushes.Transparent, // Wichtig!
-            BorderThickness = new Thickness(0),     // kein eigener Rand
-            Foreground = new SolidColorBrush(colText),
-            Padding = new Thickness(4, 4, 6, 4),
-            VerticalContentAlignment = VerticalAlignment.Center
-        };
-        _searchTb.TextChanged += (_, __) => RefreshShopSearchResults();
-        Grid.SetColumn(_searchTb, 1);
-        innerGrid.Children.Add(_searchTb);
-
-        // pack Grid in die äußere Border
-        searchOuter.Child = innerGrid;
-
-        _btnAnalyze = new Button
-        {
-            Content = " 💰 Profit Trades ",
-            Margin = new Thickness(4, 2, 4, 2),
-            Style = niceBtnStyle
-        };
-
-        _btnPathFinder = new Button
-        {
-            Content = " 🔍 Buy X for Y ",
-            Margin = new Thickness(4, 2, 4, 2),
-            Style = niceBtnStyle
-        };
-
-        //btnGo.Click += (_, __) => RefreshShopSearchResults();
-        _searchTb.TextChanged += (_, __) => RefreshShopSearchResults();
-
-        _btnAnalyze.Click += (_, __) => OpenAnalysisWindow();
-        _btnPathFinder.Click += (_, __) => OpenPathFinderWindow();
-
-
-        row1.Children.Add(searchOuter);
-      //  row1.Children.Add(btnGo);
-        row1.Children.Add(_btnAnalyze);
-        row1.Children.Add(_btnPathFinder);
-        headerWrap.Children.Add(row1);
-
-        // --- Zeile 2: Filter-Checkboxen ---
-        var row2 = new StackPanel
-        {
-            Orientation = Orientation.Horizontal,
-            Margin = new Thickness(0, 0, 0, 4)
-        };
-
-        _chkSell = new CheckBox();
-        StyleHeaderCheckBox(_chkSell, "Sells💰", true);
-        
-
-        _chkBuy = new CheckBox();
-        StyleHeaderCheckBox(_chkBuy, "Buys🛒", true);
-       
-
-        _chkHideEmpty = new CheckBox();
-        StyleHeaderCheckBox(_chkHideEmpty, "Hide 0-stock", false);
-        
-        
-        _chkNewShopAlerts = new CheckBox();
-        StyleHeaderCheckBox(_chkNewShopAlerts, "New Shops → chat", false);
-       
-
-        _chkSuspiciousAlerts = new CheckBox();
-        StyleHeaderCheckBox(_chkSuspiciousAlerts, "Suspicious → chat", false);
-        
-
-        // Events für Filter/Aktionen
-        _chkSell.Checked += (_, __) => RefreshShopSearchResults();
-        _chkSell.Unchecked += (_, __) => RefreshShopSearchResults();
-        _chkBuy.Checked += (_, __) => RefreshShopSearchResults();
-        _chkBuy.Unchecked += (_, __) => RefreshShopSearchResults();
-        _chkHideEmpty.Checked += (_, __) => RefreshShopSearchResults();
-        _chkHideEmpty.Unchecked += (_, __) => RefreshShopSearchResults();
-
-        // Events für globale Chat-Alarm-Flags
-        _chkNewShopAlerts.Checked += (_, __) => { _notifyNewShopsToChat = true; };
-        _chkNewShopAlerts.Unchecked += (_, __) => { _notifyNewShopsToChat = false; };
-
-        _chkSuspiciousAlerts.Checked += (_, __) => { _notifySuspiciousShops = true; };
-        _chkSuspiciousAlerts.Unchecked += (_, __) => { _notifySuspiciousShops = false; };
-
-        row2.Children.Add(_chkSell);
-        row2.Children.Add(_chkBuy);
-        row2.Children.Add(_chkHideEmpty);
-        row2.Children.Add(_chkNewShopAlerts);
-        row2.Children.Add(_chkSuspiciousAlerts);
-
-        headerWrap.Children.Add(row2);
-
-        // --- Zeile 3: Alerts / "+ Alert" Button + Alert-Liste ---
-        var row3 = new StackPanel
-        {
-            Orientation = Orientation.Vertical,
-            Margin = new Thickness(0, 0, 0, 4)
-        };
-
-        // Add-Alert Button
-        _btnAddAlert = new Button
-        {
-            Content = "＋ Alert (watch this)",
-            Margin = new Thickness(0, 0, 0, 4)
-        };
-
-        _btnAddAlert.Click += (_, __) => AddAlertFromCurrentSearch();
-
-        // Alert-Liste
-        _alertList = new ListBox
-        {
-            Background = SearchWinBg,
-            BorderThickness = new Thickness(1),
-            BorderBrush = new SolidColorBrush(Color.FromArgb(60, 255, 255, 255)),
-            Foreground = SearchText,
-            Height = 70
-        };
-
-        row3.Children.Add(_btnAddAlert);
-        row3.Children.Add(_alertList);
-
-        headerWrap.Children.Add(row3);
-
-        // Dock das oben rein
-        DockPanel.SetDock(headerWrap, Dock.Top);
-        root.Children.Add(headerWrap);
-
-        // PATH FINDER WINDOW
-
-
-
-        // ========== MAIN SEARCH RESULT LIST ==========
-        _searchList = new ListBox
-        {
-            Background = SearchWinBg,
-            BorderThickness = new Thickness(0),
-            Foreground = SearchText
-        };
-
-        // dezente Auswahl wie vorher
-        _searchList.ItemContainerStyle = new Style(typeof(ListBoxItem))
-        {
-            Setters =
-        {
-            new Setter(Control.BackgroundProperty, Brushes.Transparent),
-            new Setter(Control.BorderThicknessProperty, new Thickness(0)),
-            new Setter(Control.PaddingProperty, new Thickness(0))
-        }
-        };
-
-        // wir packen unten drunter noch eine Analyse-Liste (erstmal collapsed)
-        var resultWrap = new StackPanel
-        {
-            Orientation = Orientation.Vertical
-        };
-
-        resultWrap.Children.Add(_searchList);
-
-        // Separator
-        resultWrap.Children.Add(new Border
-        {
-            Height = 1,
-            Margin = new Thickness(0, 8, 0, 8),
-            Background = new SolidColorBrush(Color.FromArgb(60, 255, 255, 255))
-        });
-
-        _analysisList = new ListBox
-        {
-            Background = SearchWinBg,
-            BorderThickness = new Thickness(0),
-            Foreground = SearchText,
-            Visibility = Visibility.Collapsed // erst sichtbar machen, wenn Analyze geklickt
-        };
-
-        resultWrap.Children.Add(_analysisList);
-
-        var scroll = new ScrollViewer
-        {
-            Content = resultWrap,
-            VerticalScrollBarVisibility = ScrollBarVisibility.Auto,
-            HorizontalScrollBarVisibility = ScrollBarVisibility.Disabled
-        };
-        ApplyThinScrollbar(scroll);
-
-        // Wichtig: kein eigenes Scroll-Handling mehr hier!
-        root.Children.Add(scroll);
-
-        // Jetzt sorgen wir dafür, dass das Fenster-Mausrad an den ScrollViewer geht:
-        w.PreviewMouseWheel += (snd, e) =>
-        {
-            // nur eingreifen, wenn der Cursor irgendwo im Scroll-Bereich ist
-            // (damit du oben in den Controls nicht plötzlich scrollst, wenn du eigentlich Text selektierst)
-            if (!scroll.IsMouseOver) return;
-
-            // standardmäßiges Verhalten: 1 "Wheel-Delta" (120) = ~3-4 Zeilen
-            // Wir machen kleine Schritte, nicht riesige Sprünge.
-            const double step = 30.0; // Pixel pro Wheel Tick, fühlt sich meist gut an
-            double direction = e.Delta > 0 ? -1 : 1;
-
-            double newOffset = scroll.VerticalOffset + direction * step;
-            if (newOffset < 0) newOffset = 0;
-            if (newOffset > scroll.ScrollableHeight) newOffset = scroll.ScrollableHeight;
-
-            scroll.ScrollToVerticalOffset(newOffset);
-            e.Handled = true;
-        };
-
-        w.Content = root;
-        _shopSearchWin = w;
-
-        // wenn Fenster geschlossen wird, Referenz löschen
-        _shopSearchWin.Closed += (_, __) => _shopSearchWin = null;
-
-        // ganz am Ende: initial einmal UI auffüllen
-        // nur laden, wenn wir noch keine saved alerts drin haben
-        if (_alertRules.All(r => !r.IsSaved))
-        {
-            LoadPersistentAlerts();
-        }
-        RefreshAlertListUI();
-        RefreshShopSearchResults();
-    }
 
     // PATH FINDER WINDOW LOGIK
 
@@ -9282,53 +6579,6 @@ public partial class MainWindow : Window
         }
     }
 
-    private void RefreshShopSearchResults()
-    {
-        if (_shopSearchWin == null || _searchList == null) return;
-
-        string q = _searchTb?.Text?.Trim() ?? "";
-
-        bool wantSell = _chkSell?.IsChecked != false;
-        bool wantBuy = _chkBuy?.IsChecked != false;
-
-        bool hideEmpty = _chkHideEmpty?.IsChecked == true;
-
-        // Helper zum Text-Match
-        bool MatchesLeft(RustPlusClientReal.ShopOrder o)
-            => string.IsNullOrEmpty(q)
-               || ResolveItemName(o.ItemId, o.ItemShortName)
-                  .Contains(q, StringComparison.OrdinalIgnoreCase);
-
-        bool MatchesRight(RustPlusClientReal.ShopOrder o)
-            => string.IsNullOrEmpty(q)
-               || ResolveItemName(o.CurrencyItemId, o.CurrencyShortName)
-                  .Contains(q, StringComparison.OrdinalIgnoreCase);
-
-        _searchList.Items.Clear();
-
-        foreach (var s in _lastShops)
-        {
-            if (s.Orders == null || s.Orders.Count == 0) continue;
-
-            // Deine angepasste Filter-Logik:
-            var offers = s.Orders
-                .Where(o =>
-                    ((wantSell && MatchesLeft(o)) || (wantBuy && MatchesRight(o))) &&
-                    (!hideEmpty || o.Stock > 0)
-                )
-                .ToList();
-
-            if (hideEmpty && offers.Count == 0)
-                continue; // ganzer Shop raus, weil nur 0-stock
-
-            if (offers.Count == 0)
-                continue;
-
-            // Karte bauen und hinzufügen
-            _searchList.Items.Add(BuildShopSearchCard(s, offers, compact: false));
-        }
-    }
-
     private FrameworkElement BuildSearchResultCard(
     RustPlusClientReal.ShopMarker shop,
     IEnumerable<RustPlusClientReal.ShopOrder> offers)
@@ -9380,43 +6630,6 @@ public partial class MainWindow : Window
         };
 
         return border;
-    }
-
-    // Weltpunkt (x,y) in die Mitte des sichtbaren Bereichs schieben – Zoom bleibt unverändert
-    private void CenterMapOnWorld(double x, double y)
-    {
-        if (_worldSizeS <= 0) return;
-
-        // 1) Welt -> Szenenpixel (Bildkoordinaten innerhalb des Welt-Quadrats)
-        var p = WorldToImagePx(x, y); // benutzt _worldRectPx + _worldSizeS
-
-        // 2) Viewbox-Skalierung + Offsets (Uniform-Scale + evtl. Letterbox-Ränder)
-        var (s, offX, offY) = GetViewboxScaleAndOffset(); // <- hast du schon (wird auch fürs Panning benutzt)
-
-        // 3) Gewünschte Host-Mitte
-        double hostCx = WebViewHost.ActualWidth * 0.5;
-        double hostCy = WebViewHost.ActualHeight * 0.5;
-
-        // 4) Aktuelle Matrix lesen, nur die Translation neu setzen
-        var m = MapTransform.Matrix;
-        double sx = Math.Abs(m.M11) < 1e-9 ? 1 : m.M11;
-        double sy = Math.Abs(m.M22) < 1e-9 ? 1 : m.M22;
-
-        // Wir wollen: off + s * (scale * p + offset) == HostCenter  ⇒  offset = (HostCenter - off)/s - scale*p
-        m.OffsetX = (hostCx - offX) / s - sx * p.X;
-        m.OffsetY = (hostCy - offY) / s - sy * p.Y;
-
-        MapTransform.Matrix = m;
-    }
-
-    private void ShopElement_Click(object sender, MouseButtonEventArgs e)
-    {
-        if (sender is FrameworkElement fe && fe.Tag is RustPlusClientReal.ShopMarker s)
-        {
-            CenterMapOnWorld(s.X, s.Y);
-            AppendLog($"Shop clicked: {s.Label ?? "(no Label)"} @ ({s.X:0},{s.Y:0}) | offers={s.Orders?.Count ?? 0}");
-            // hier könntest du später ein richtiges Popup öffnen
-        }
     }
 
     // SHOP ANALYTICS AND ALARM MECHANICS
@@ -9688,13 +6901,6 @@ public partial class MainWindow : Window
 
     // ====== SHOP SEARCH WINDOW UI-Elemente ======
     // Erweiterungen, die wir neu brauchen:
-    private CheckBox? _chkHideEmpty;            // "Hide 0-stock"
-    private CheckBox? _chkNewShopAlerts;        // "New shop alerts to chat"
-    private CheckBox? _chkSuspiciousAlerts;     // "Suspicious shop alerts"
-    private Button? _btnAddAlert;               // "+ Alert" Button
-    private ListBox? _alertList;                // Liste der aktiven Alerts im UI
-    private Button? _btnAnalyze;                // "Analyze" Button
-    private Button? _btnPathFinder;                // "PathFinder" Button
     private ListBox? _analysisList;             // Ergebnisse der Analyse
 
     private DateTime _initialShopSnapshotTime = DateTime.UtcNow; // set beim allerersten erfolgreichen Poll
@@ -9762,6 +6968,9 @@ public partial class MainWindow : Window
         pillButtonStyle.Setters.Add(new Setter(Control.CursorProperty, Cursors.Hand));
         // CornerRadius geht nur über ControlTemplate hacky;
         // Quick&dirty ohne Template: wir lassen’s rechteckig mit 4er Radius über Border below:
+
+        // Push to WebView2 shop search panel (replaces WPF _alertList when window is open)
+        _ = PushAlertsToWebViewAsync();
 
         if (_alertList == null) return;
 
@@ -11129,6 +8338,38 @@ public partial class MainWindow : Window
     {
         var modal = new SettingsModal { Owner = this };
         modal.ShowDialog();
+        ApplySettings();
+    }
+    
+    private void ApplySettings()
+    {
+        if (TxtLog != null)
+        {
+            TxtLog.Visibility = TrackingService.HideConsole ? Visibility.Collapsed : Visibility.Visible;
+        }
+
+        if (ColSidebar != null)
+        {
+            double w = TrackingService.SidebarWidth;
+            if (w < 400) w = 400;
+            
+            // Ensure we don't squash the map below 800 if window is small
+            if (this.ActualWidth > 0)
+            {
+                double maxW = this.ActualWidth - 850; // 800 map + 50 padding/splitter
+                if (w > maxW && maxW > 400) w = maxW;
+            }
+
+            ColSidebar.Width = new GridLength(w, GridUnitType.Pixel);
+        }
+    }
+
+    private void SidebarSplitter_DragCompleted(object sender, System.Windows.Controls.Primitives.DragCompletedEventArgs e)
+    {
+        if (ColSidebar != null)
+        {
+            TrackingService.SidebarWidth = ColSidebar.ActualWidth;
+        }
     }
     private async void BtnCheckUpdates_Click(object sender, RoutedEventArgs e)
     {
